==> 3_programs/6_soumission.R <==
## Aline Davias
## Analyses microbiot et neurodeveloppement 
## 20/02/2024


# Packages loading ----
library(tidyverse)
library(phyloseq)
library(expss)
library(gtsummary)
library(labelled)
library(questionr)
library(sjlabelled)
library(openxlsx)
library(corrplot)
library(see)
library(psych)
library(compositions)
library(writexl)
library(performance)
library(see)
library(broom)
library(broom.helpers)
library(broom.mixed)
library(mice)
library(gt)
library(GGally)
library(ggrepel)
library(performance)
library(effectsize)
theme_gtsummary_compact()

# Data and functions loading ----
load("1_intermediate_data/4_data_standardization_AD_gumme.RData")
rm(resultats_correlation, bdd_final_imp, bdd_final_imp_sensi_seuil, 
   bdd_std_large, bdd_std_long)
source("3_programs/5_functions_AD_gumme.R")
corres <- 
  taxa_table_Y1 %>% 
  select(Phyla_corres = ch_feces_phylum_ASVbased_Y1, 
         Class_corres = ch_feces_class_ASVbased_Y1,
         Order_corres = ch_feces_order_ASVbased_Y1, 
         Family_corres = ch_feces_family_ASVbased_Y1,
         Genera_corres = ch_feces_genus_ASVbased_Y1) %>%
  filter(Genera_corres %in% genera_var_labels) %>%
  distinct(Genera_corres, .keep_all = TRUE)

# Vectors ----
covariates <- covar_vec_model_final
explanatory <- bdd_final_imp_1 %>% 
  select(all_of(microbiote_std_vec)) %>% 
  select(-ch_feces_SpecRich_10000_ASV_std_Y1_10, # pour analyses de sensibilité
         -ch_feces_Shannon_10000_ASV_std_Y1) %>% # pour analyses de sensibilité) 
  colnames()

outcomes <- bdd %>%
  select(all_of(neuro_vec)) %>%
  select(-ch_socawar_y3,
         -ch_soccog_y3,
         -ch_soccom_y3, 
         -ch_socmot_y3, 
         -ch_RRB_y3) %>%
  colnames()
rm(neuro_vec, microbiote_vec, covar_a_imputer, covar_a_tester, covar_vec_model_final, microbiote_std_vec)

genera_var_names <- genera_var_names %>%
  str_replace_all("imp_log_Y1", "imp_log_std_Y1")
genera_var_names_raw <- genera_var_names %>%
  str_replace_all("_imp_log_std_Y1", "_Y1")

test <- bdd %>% select(all_of(genera_var_names_raw))                     # récupération des labels des variables
colnames(test) <- gsub("_Y1", "_imp_log_std_Y1", colnames(test))
test <- test %>% 
  var_label() %>% 
  as.data.frame() %>% 
  t() %>% 
  as.matrix() %>% 
  as.data.frame() %>% 
  rownames_to_column(var = "Exposure") %>%
  mutate(Genera_corres = gsub("One year child feces relative abundance of genus ", "", V1)) %>%
  select(-V1)
corres <-                                                                         # Ajout des correspondances taxonomiques 
  left_join(corres, test, by = "Genera_corres")
rm(test)

# Data description ----
## Covariates ----
descrip_covar <- 
  bdd_final_imp_1 %>% 
  select(all_of(covariates)) %>% 
  tbl_summary(type = list(all_categorical() ~ "categorical"), 
              label = list(ch_care_main_12m_opt2_2c ~ "Main mode of child care at 12 months")) %>%
  bold_labels()

descrip_covar

## Included genera (n=46) ----
### Tables ----
descrip_genera <- tbl_merge(
  tbls = 
    list(
      tbl_1 = 
        bdd %>% 
        select(all_of(genera_var_names_raw)) %>%
        na.omit()%>%
        set_label(genera_var_labels) %>%
        tbl_summary(
          type = list(everything() ~ "continuous"), 
          statistic = list(everything() ~ "{median} ({p33}, {p66})"), 
          digits = list(all_continuous() ~ c(2, 3, 3))), 
      tbl_2 = 
        bdd_final_imp_1 %>% 
        select(all_of(genera_var_names)) %>%
        rename_all(~gsub("imp_log_std_", "", .)) %>%
        set_label(genera_var_labels) %>%
        tbl_summary(
          type = list(everything() ~ "continuous"), 
          statistic = list(everything() ~ "{median} ({p25}, {p75})"), 
          digits = list(all_continuous() ~ c(2, 1, 1))), 
      tbl_3 = 
        bdd %>% 
        select(all_of(genera_var_names_raw)) %>%
        na.omit()%>%
        mutate_all(~ ifelse(.>0, "Yes", "No")) %>%
        set_label(genera_var_labels) %>%
        tbl_summary(
          type = list(everything() ~ "categorical"))), 
  tab_spanner = c("**Continuous**", "**Log transformed**", "**Categorical (Y/N)**"))
descrip_genera

### Density plots ----
#### Distribution avant imputation et transformation 
densityplot(data = bdd, vars = genera_var_names_raw[1:23], ncol = 5)    
densityplot(data = bdd, vars = genera_var_names_raw[24:46], ncol = 5)

#### Distribution après imputation et transformation 
densityplot(data = bdd_final_imp_1, vars = genera_var_names[1:23], ncol = 5)    
densityplot(data = bdd_final_imp_1, vars = genera_var_names[24:46], ncol = 5)


# Calcul du nombre de tests effectifs pour la correction pour comp multiple (FWER) ----
test_expo_taxa <- bdd_final_imp_1 %>%
  filter(!is.na(ch_feces_rel_p1_std_Y1_10)) %>%
  select(all_of(explanatory)) %>%
  select(-"ch_feces_SpecRich_5000_ASV_std_Y1_10", -"ch_feces_Shannon_5000_ASV_std_Y1") 
test_expo_taxa <- lapply(test_expo_taxa, function(x) { var_lab(x) <- NULL; return(x) })
test_expo_taxa <- as.data.frame(test_expo_taxa)
cor_mixed_table_expo_taxa <- cor_mixed(data = test_expo_taxa)
results_alpha_corrected_expo_taxa <- alpha_corrected(data = test_expo_taxa, alpha = 0.05)
results_M0_corrected_expo_taxa <- M0_corrected(data = test_expo_taxa, alpha = 0.05)
results_M0_corrected_expo_taxa

test_expo_outcomes <- bdd_final_imp_1 %>%
  select(all_of(outcomes))
test_expo_outcomes <- lapply(test_expo_outcomes, function(x) { var_lab(x) <- NULL; return(x) })
test_expo_outcomes <- as.data.frame(test_expo_outcomes)
cor_mixed_table_expo_outcomes <- cor_mixed(data = test_expo_outcomes)
results_alpha_corrected_expo_outcomes <- alpha_corrected(data = test_expo_outcomes, alpha = 0.05)
results_M0_corrected_expo_outcomes <- M0_corrected(data = test_expo_outcomes, alpha = 0.05)
results_M0_corrected_expo_outcomes

## Pour les analyses de diversité, le seuil corrigé est : 0.05/(3*7)
## Pour les analyses de taxonomie, le seuil corrigé est : 0.05/(33*7)
rm(test_expo_taxa, test_expo_outcomes, 
   cor_mixed_table_expo_taxa, cor_mixed_table_expo_outcomes, 
   results_alpha_corrected_expo_taxa, results_alpha_corrected_expo_outcomes,
   results_M0_corrected_expo_taxa, results_M0_corrected_expo_outcomes, 
   cor_mixed, alpha_corrected, M0_corrected)

# Running linear regressions ----
## adapation des covariables selon l'outcome
covariates_CBCL <- c("ch_age_CBCL_Y2", covariates)                    
covariates_SRS_BRIEF <- c("ch_age_SRS_BRIEFP_Y3", covariates)
covariates_IQ <- covariates   # les variables WPPSI sont déjà standardisées sur l'âge et le neuropsychologue

covariates_map <- list(
  CBCL = covariates_CBCL,
  SRS_BRIEF = covariates_SRS_BRIEF,
  IQ = covariates_IQ
)

## Création d'une liste pour stocker les tableaux par outcome
tbls_by_outcome_multi <- vector("list", length(outcomes))
names(tbls_by_outcome_multi) <- outcomes

# Boucle principale
for (outcome in outcomes) {
  # Sélection des covariables appropriées
  if (outcome %in% c("ch_cbclintscore_y2", "ch_cbclextscore_y2")) {
    covars <- covariates_map$CBCL
  } else if (outcome %in% c("ch_SRStotal_y3", "ch_briefpinhibit_y3", "ch_briefpshift_y3", "ch_briefpemocontrol_y3", "ch_briefpworkmemo_y3", "ch_briefpplan_y3")) {
    covars <- covariates_map$SRS_BRIEF
  } else if (outcome %in% c("ch_WPPSI_verbal_comprehension_cor_Y3", "ch_WPPSI_visuospatiale_cor_Y3", "ch_WPPSI_work_memory_cor_Y3", "ch_WPPSI_total_cor_Y3")) {
    covars <- covariates_map$IQ
  }
  
  tbls_for_outcome_multi <- vector("list", length(explanatory))
  names(tbls_for_outcome_multi) <- explanatory
  
  for (exposure in explanatory) {                                               # running linear regression
    terms <- c(exposure, covars)
    formula <- reformulate(terms, response = outcome)
    model <- lm(formula, data = bdd_final_imp_1)
    # model <- with(data = bdd_final_imp, 
    #               exp = lm(formula))
    
    tbl <-                                                                      
      tbl_regression(
        model, 
        include = exposure,
        estimate_fun = scales::label_number(accuracy = .01, decimal.mark = "."),
        pvalue_fun = custom_pvalue_fun,
        exponentiate = FALSE) %>%
      bold_p() %>%
      bold_labels() %>%
      add_global_p(include = exposure, singular.ok = TRUE, keep = TRUE)
    
    tbls_for_outcome_multi[[exposure]] <- tbl
  }
  tbls_by_outcome_multi[[outcome]] <- tbls_for_outcome_multi
}

table_multi <- tibble()                                                         # Initialisation d'un tibble vide pour stocker les résultats finaux
for (i in seq_along(tbls_by_outcome_multi)) {                                   # Nom de l'outcome pour cette itération
  outcome_name <- names(tbls_by_outcome_multi)[i]
  
  for (j in seq_along(tbls_by_outcome_multi[[i]])) {                            # Itération sur chaque tbl_regression dans la liste courante
    exposure_name <- names(tbls_by_outcome_multi[[i]])[j]                       # Nom de la variable d'exposition pour cette itération
    tbl_data <- tbls_by_outcome_multi[[i]][[j]] %>%                             # Extraction des données du tableau tbl_regression
      as_tibble() %>%
      mutate(Outcome = outcome_name, Exposure = exposure_name)
    
    table_multi <- bind_rows(table_multi, tbl_data)                             # Ajout des données extraites au tibble final
  }
}
rm(terms, formula, model,
   tbl, tbl_data,
   tbls_for_outcome_multi,
   exposure, exposure_name, i, j, outcome, outcome_name, 
   covars)


table_multi <-                                                                  # Ajout des correspondances taxonomiques 
  left_join(table_multi, corres, by = "Exposure") %>%
  select(Phyla_corres, Class_corres, Order_corres, Family_corres, Genera_corres, everything())
rm(corres)

table_multi <- table_multi %>%
  select(Outcome, 
         Phyla_corres, Class_corres, Order_corres, Family_corres, Genera_corres,
         Exposure, 
         Beta = "**Beta**", 
         "95% CI" = "**95% CI**",
         "p-value" = "**p-value**", 
         "Characteristic" = "**Characteristic**") %>%
  mutate(
    `Gut microbiota parameters` = ifelse(!is.na(Genera_corres), Genera_corres, Exposure),
    `Gut microbiota parameters` = 
      fct_recode(`Gut microbiota parameters`, 
                 "Firmicutes" = "ch_feces_rel_p1_std_Y1_10",
                 "Actinobacteria" = "ch_feces_rel_p2_std_Y1_10",
                 "Bacteroidetes" = "ch_feces_rel_p3_std_Y1_10",
                 "Proteobacteria" = "ch_feces_rel_p4_std_Y1_10",
                 "Shannon diversity" = "ch_feces_Shannon_5000_ASV_std_Y1",
                 "Specific richness" = "ch_feces_SpecRich_5000_ASV_std_Y1_10",
                 "Clostridium IV" = "Clostridium_IV",
                  "Clostridium sensu stricto" = "Clostridium_sensu_stricto",
                  "Clostridium XlVa" = "Clostridium_XlVa",
                  "Clostridium XVIII" = "Clostridium_XVIII",
                  "Erysipelotrichaceae incertae sedis" = "Erysipelotrichaceae_incertae_sedis",
                  "Escherichia and Shigella" = "Escherichia_Shigella",
                  "Lachnospiracea incertae sedis" = "Lachnospiracea_incertae_sedis",
                  "Ruminococcus 2" = "Ruminococcus2",
                  "Saccharibacteria genera incertae sedis" = "Saccharibacteria_genera_incertae_sedis"),
    `Gut microbiota parameters` = 
      fct_relevel(`Gut microbiota parameters`, 
                  "Saccharibacteria genera incertae sedis",
                  "Peptoniphilus", "Granulicatella", "Anaerotruncus", "Lactococcus", 
                  "Terrisporobacter", "Oscillibacter", "Haemophilus",
                  "Coprococcus", "Erysipelotrichaceae incertae sedis", "Butyricicoccus", 
                  "Dialister",  "Subdoligranulum", "Intestinibacter", "Klebsiella",
                  "Eisenbergiella", "Hungatella","Dorea", "Eggerthella",
                  "Romboutsia","Clostridium IV","Ruminococcus 2","Flavonifractor", "Alistipes",
                  "Collinsella","Parabacteroides","Fusicatenibacter", "Veillonella",
                  "Roseburia", "Enterobacter","Cellulosibacter",  "Enterococcus",
                  "Clostridium sensu stricto", "Clostridium XVIII", "Ruminococcus", 
                   "Gemmiger", "Anaerostipes", "Lachnospiracea incertae sedis", 
                  "Clostridium XlVa",  "Streptococcus", "Faecalibacterium", "Akkermansia",
                  "Escherichia and Shigella", "Blautia", "Bacteroides","Bifidobacterium", 
                  "Proteobacteria", "Bacteroidetes","Actinobacteria", "Firmicutes",
                  "Shannon diversity", "Specific richness"),
    Outcome = 
      fct_recode(Outcome, 
                 "Emotional control BRIEF-P sub-score at 3 years" = "ch_briefpemocontrol_y3",
                  "Inhibition BRIEF-P sub-score at 3 years" = "ch_briefpinhibit_y3",
                  "Plan and organization BRIEF-P sub-score at 3 years" = "ch_briefpplan_y3",
                  "Shift BRIEF-P sub-score at 3 years" = "ch_briefpshift_y3",
                  "Work memory BRIEF-P sub-score at 3 years" = "ch_briefpworkmemo_y3",
                  "Externalizing CBCL sub-score at 2 years" = "ch_cbclextscore_y2",
                  "Internalizing CBCL sub-score at 2 years" = "ch_cbclintscore_y2",
                  "Total SRS score at 3 years" = "ch_SRStotal_y3",
                  "Total WPPSI score at 3 years" = "ch_WPPSI_total_cor_Y3",
                  "Verbal comprehension WPPSI sub-score at 3 years" = "ch_WPPSI_verbal_comprehension_cor_Y3",
                  "Visuospatial WPPSI sub-score at 3 years" = "ch_WPPSI_visuospatiale_cor_Y3",
                  "Work memory WPPSI sub-score at 3 years" = "ch_WPPSI_work_memory_cor_Y3"),
    Outcome = 
      fct_relevel(Outcome,
                  "Internalizing CBCL sub-score at 2 years", "Externalizing CBCL sub-score at 2 years",
                  "Emotional control BRIEF-P sub-score at 3 years", "Inhibition BRIEF-P sub-score at 3 years",
                  "Plan and organization BRIEF-P sub-score at 3 years", "Shift BRIEF-P sub-score at 3 years",
                  "Work memory BRIEF-P sub-score at 3 years", "Total SRS score at 3 years",
                  "Verbal comprehension WPPSI sub-score at 3 years", "Visuospatial WPPSI sub-score at 3 years",
                  "Work memory WPPSI sub-score at 3 years", "Total WPPSI score at 3 years"),
    Outcome_rec = gsub("sub-score at 3 years", "Y3", Outcome), 
    Outcome_rec = gsub("sub-score at 2 years", "Y2", Outcome_rec), 
    Outcome_rec = gsub("score at 3 years", "Y3", Outcome_rec), 
    `p-value` = gsub("__", "", `p-value`),
    `p-value` = as.numeric(`p-value`),
    `q-value` = case_when(
      `Gut microbiota parameters` == "Specific richness" ~ `p-value`/(3*7),     # correction spéciale analyses de diversité 
      `Gut microbiota parameters` == "Shannon diversity" ~ `p-value`/(3*7),     # correction spéciale analyses de diversité 
      .default = `p-value`/(33*7)),                                             # correction spéciale analyses de taxonomie 
    p_value_shape = ifelse(`p-value`<0.05, "p-value<0.05", "p-value≥0.05"),
    q_value_shape = ifelse(`q-value`<0.05, "q-value<0.05", "q-value≥0.05"), 
    sens_beta = ifelse(Beta < 0, "Beta<0", "Beta≥0"), 
    sens_beta = fct_relevel(sens_beta, "Beta≥0", "Beta<0"), 
    improved_neuro = 
      case_when(Outcome %in% c("Verbal comprehension WPPSI sub-score at 3 years",
                               "Visuospatial WPPSI sub-score at 3 years", 
                               "Work memory WPPSI sub-score at 3 years", 
                               "Total WPPSI score at 3 years") & Beta>0 ~ "Improved neurodevelopmental outcome", 
                Outcome %in% c("Verbal comprehension WPPSI sub-score at 3 years",
                               "Visuospatial WPPSI sub-score at 3 years", 
                               "Work memory WPPSI sub-score at 3 years", 
                               "Total WPPSI score at 3 years") & Beta<0 ~ "Impaired neurodevelopmental outcome", 
                Outcome %in% c("Internalizing CBCL sub-score at 2 years", 
                               "Externalizing CBCL sub-score at 2 years", 
                               "Total SRS score at 3 years", 
                               "Inhibition BRIEF-P sub-score at 3 years", 
                               "Shift BRIEF-P sub-score at 3 years", 
                               "Emotional control BRIEF-P sub-score at 3 years", 
                               "Work memory BRIEF-P sub-score at 3 years", 
                               "Plan and organization BRIEF-P sub-score at 3 years") & Beta>0 ~ "Impaired neurodevelopmental outcome",
                Outcome %in% c("Internalizing CBCL sub-score at 2 years", 
                               "Externalizing CBCL sub-score at 2 years", 
                               "Total SRS score at 3 years", 
                               "Inhibition BRIEF-P sub-score at 3 years", 
                               "Shift BRIEF-P sub-score at 3 years", 
                               "Emotional control BRIEF-P sub-score at 3 years", 
                               "Work memory BRIEF-P sub-score at 3 years", 
                               "Plan and organization BRIEF-P sub-score at 3 years") & Beta<0 ~ "Improved neurodevelopmental outcome"))  %>% 
  separate(col = "95% CI", into = c("lower_CI", "upper_CI"), sep = ",", remove = FALSE) %>%
  mutate(
    lower_CI = as.numeric(lower_CI),
    upper_CI = as.numeric(upper_CI)
  ) %>%
  select(
    Phyla_corres, Class_corres, Order_corres, Family_corres,
    Outcome, Outcome_rec, 
    `Gut microbiota parameters`,
    Exposure, 
    Beta, sens_beta, 
    "95% CI", lower_CI, upper_CI, 
    "p-value", p_value_shape, 
    "q-value", q_value_shape, 
    improved_neuro) 

# Tables ----
## Table 1: Covariables description ----
table_1 <- bdd %>% 
  select(all_of(covariates), 
         all_of(covar_age), 
         ch_feces_rel_p1_Y1) %>%
  mutate(
    statu = ifelse(!is.na(ch_feces_rel_p1_Y1), "included", "excluded"), 
    statu = fct_relevel(statu,  "included", "excluded")) %>%
  select(-ch_feces_rel_p1_Y1) %>%
  tbl_summary(
    type = list(all_categorical() ~ "categorical"), 
    label = list(ch_care_main_12m_opt2_2c ~ "Main mode of child care at 12 months"), 
    by="statu", 
    missing = "no", 
    digits = list(mo_age ~ 0, 
                  po_gd ~ 0, 
                  home_total_y3 ~ 0, 
                  mo_hadtotscore_grt3_imp ~ 0, 
                  ch_age_CBCL_Y2 ~ 0, 
                  ch_age_SRS_BRIEFP_Y3 ~ 0)) %>%
  add_p(pvalue_fun = function(x) formatC(x, format = "f", digits = 3)) %>%
  bold_labels() %>%
  add_n()

## Table 2: Associations alphadiv and neurodev ----
table_2_long <-
  tbl_stack(tbls =
              lapply(tbls_by_outcome_multi, function(tbl_list) {
                tbl_merge(
                  list(
                    tbl_list$ch_feces_SpecRich_5000_ASV_std_Y1_10 %>% add_n(),   
                    tbl_list$ch_feces_Shannon_5000_ASV_std_Y1 %>% add_n()),  
                  tab_spanner = c("**Specific richness**", "**Shannon diversity**"))}),
            group_header = spanner_names)

table_2_large <- 
  tbl_stack(tbls = list(
    tbl_merge(
      tbls = lapply(1:length(tbls_by_outcome_multi), function(i) tbls_by_outcome_multi[[i]]$ch_feces_SpecRich_5000_ASV_std_Y1_10),  
      tab_spanner = spanner_names),
    tbl_merge( 
      tbls = lapply(1:length(tbls_by_outcome_multi), function(i) tbls_by_outcome_multi[[i]]$ch_feces_Shannon_5000_ASV_std_Y1),  
      tab_spanner = spanner_names)))

table_2 <- list(table_2_long = table_2_long, 
                table_2_large = table_2_large)
rm(table_2_long, table_2_large)

## Table 3: Associations betadiv and neurodev ----
# cf code spécifique

## Table 4: Associations phyla and neurodev ----
table_4_long <-
  tbl_stack(tbls =
              lapply(tbls_by_outcome_multi, function(tbl_list) {
                tbl_merge(
                  list(tbl_list$ch_feces_rel_p1_std_Y1_10 %>% add_n(),   # correspond Firmicutes
                       tbl_list$ch_feces_rel_p2_std_Y1_10 %>% add_n(),   # correspond Actinobacteria
                       tbl_list$ch_feces_rel_p3_std_Y1_10 %>% add_n(),   # correspond Bacteroidetes
                       tbl_list$ch_feces_rel_p4_std_Y1_10 %>% add_n()),  # correspond Proteobacteria
                  tab_spanner = c("**Firmicutes**", 
                                  "**Actinobacteria**", 
                                  "**Bacteroidetes**", 
                                  "**Proteobacteria**"))}),
            group_header = spanner_names)

table_4_large <- 
  tbl_stack(tbls = list(
    tbl_merge(
      tbls = lapply(1:length(tbls_by_outcome_multi), function(i) tbls_by_outcome_multi[[i]]$ch_feces_rel_p1_std_Y1_10),  # correspond à Firmicutes
      tab_spanner = spanner_names),
    tbl_merge(
      tbls = lapply(1:length(tbls_by_outcome_multi), function(i) tbls_by_outcome_multi[[i]]$ch_feces_rel_p2_std_Y1_10),  # correspond à Actinobacteria
      tab_spanner = spanner_names),
    tbl_merge(
      tbls = lapply(1:length(tbls_by_outcome_multi), function(i) tbls_by_outcome_multi[[i]]$ch_feces_rel_p3_std_Y1_10),  # correspond à Bacteroidetes
      tab_spanner = spanner_names),
    tbl_merge(
      tbls = lapply(1:length(tbls_by_outcome_multi), function(i) tbls_by_outcome_multi[[i]]$ch_feces_rel_p4_std_Y1_10),  # correspond à Proteobacteria
      tab_spanner = spanner_names)))

table_4 <- list(table_4_long = table_4_long, 
                table_4_large = table_4_large)
rm(table_4_long, table_4_large)



## Table 5 ----
## ch_cbclintscore_y2 
results %>% filter(p.value<0.05) %>% filter(outcome == "ch_cbclintscore_y2")%>% select(term)
model_reduit_cbcl_int <- lm(
  ch_cbclintscore_y2 ~
    ch_age_CBCL_Y2 + 
    po_w_kg_3cat + 
    po_he_3cat + 
    mo_dipl_2cat + 
    mo_age + 
    mo_bmi_bepr_3cat + 
    ch_sex + 
    mo_par_2cat +
    ch_bf_duration_till48w_4cat + 
    po_gd + 
    po_delmod +
    ch_food_intro_Y1_3cat + 
    mo_pets + 
    ch_antibio_Y1_2cat + 
    home_total_y3 + 
    mo_hadtotscore_grt3_imp + 
    mo_tob_gr_anyt_yn_n2 +
    ch_tabacco_passive_up_to_Y1 + 
    ch_care_main_12m_opt2_2c,
  data = bdd_final_imp_1)

model_complet_cbcl_int <- lm(
  ch_cbclintscore_y2 ~
    ch_feces_rel_g19_imp_log_std_Y1 +
    ch_feces_rel_g21_imp_log_std_Y1 +
    ch_feces_rel_g36_imp_log_std_Y1 +
    ch_feces_rel_g45_imp_log_std_Y1 +
    ch_age_CBCL_Y2 + 
    po_w_kg_3cat + 
    po_he_3cat + 
    mo_dipl_2cat + 
    mo_age + 
    mo_bmi_bepr_3cat + 
    ch_sex + 
    mo_par_2cat +
    ch_bf_duration_till48w_4cat + 
    po_gd + 
    po_delmod +
    ch_food_intro_Y1_3cat + 
    mo_pets + 
    ch_antibio_Y1_2cat + 
    home_total_y3 + 
    mo_hadtotscore_grt3_imp + 
    mo_tob_gr_anyt_yn_n2 +
    ch_tabacco_passive_up_to_Y1 + 
    ch_care_main_12m_opt2_2c,
  data = bdd_final_imp_1)

## ch_cbclextscore_y2 
results %>% filter(p.value<0.05) %>% filter(outcome == "ch_cbclextscore_y2") %>% select(term)

model_reduit_cbcl_ext <- lm(
  ch_cbclextscore_y2 ~
    ch_age_CBCL_Y2 + 
    po_w_kg_3cat + 
    po_he_3cat + 
    mo_dipl_2cat + 
    mo_age + 
    mo_bmi_bepr_3cat + 
    ch_sex + 
    mo_par_2cat +
    ch_bf_duration_till48w_4cat + 
    po_gd + 
    po_delmod +
    ch_food_intro_Y1_3cat + 
    mo_pets + 
    ch_antibio_Y1_2cat + 
    home_total_y3 + 
    mo_hadtotscore_grt3_imp + 
    mo_tob_gr_anyt_yn_n2 +
    ch_tabacco_passive_up_to_Y1 + 
    ch_care_main_12m_opt2_2c,
  data = bdd_final_imp_1)

model_complet_cbcl_ext <- lm(
  ch_cbclextscore_y2 ~
    ch_feces_rel_p4_std_Y1_10 +
    ch_feces_rel_g1_imp_log_std_Y1 +
    ch_feces_rel_g13_imp_log_std_Y1 +
    ch_feces_rel_g15_imp_log_std_Y1 +
    ch_feces_rel_g19_imp_log_std_Y1 +
    ch_feces_rel_g34_imp_log_std_Y1 +
    ch_feces_rel_g39_imp_log_std_Y1 +
    ch_feces_rel_g43_imp_log_std_Y1 +
    ch_age_CBCL_Y2 + 
    po_w_kg_3cat + 
    po_he_3cat + 
    mo_dipl_2cat + 
    mo_age + 
    mo_bmi_bepr_3cat + 
    ch_sex + 
    mo_par_2cat +
    ch_bf_duration_till48w_4cat + 
    po_gd + 
    po_delmod +
    ch_food_intro_Y1_3cat + 
    mo_pets + 
    ch_antibio_Y1_2cat + 
    home_total_y3 + 
    mo_hadtotscore_grt3_imp + 
    mo_tob_gr_anyt_yn_n2 +
    ch_tabacco_passive_up_to_Y1 + 
    ch_care_main_12m_opt2_2c,
  data = bdd_final_imp_1)



## ch_SRStotal_y3
results %>% filter(p.value<0.05) %>% filter(outcome == "ch_SRStotal_y3") %>% select(term)

model_reduit_srs <- lm(
  ch_SRStotal_y3 ~
    ch_age_SRS_BRIEFP_Y3 + 
    po_w_kg_3cat + 
    po_he_3cat + 
    mo_dipl_2cat + 
    mo_age + 
    mo_bmi_bepr_3cat + 
    ch_sex + 
    mo_par_2cat +
    ch_bf_duration_till48w_4cat + 
    po_gd + 
    po_delmod +
    ch_food_intro_Y1_3cat + 
    mo_pets + 
    ch_antibio_Y1_2cat + 
    home_total_y3 + 
    mo_hadtotscore_grt3_imp + 
    mo_tob_gr_anyt_yn_n2 +
    ch_tabacco_passive_up_to_Y1 + 
    ch_care_main_12m_opt2_2c,
  data = bdd_final_imp_1)

model_complet_srs <- lm(
  ch_SRStotal_y3 ~
    ch_feces_rel_p4_std_Y1_10 +
    ch_age_SRS_BRIEFP_Y3 + 
    po_w_kg_3cat + 
    po_he_3cat + 
    mo_dipl_2cat + 
    mo_age + 
    mo_bmi_bepr_3cat + 
    ch_sex + 
    mo_par_2cat +
    ch_bf_duration_till48w_4cat + 
    po_gd + 
    po_delmod +
    ch_food_intro_Y1_3cat + 
    mo_pets + 
    ch_antibio_Y1_2cat + 
    home_total_y3 + 
    mo_hadtotscore_grt3_imp + 
    mo_tob_gr_anyt_yn_n2 +
    ch_tabacco_passive_up_to_Y1 + 
    ch_care_main_12m_opt2_2c,
  data = bdd_final_imp_1)

## ch_briefpinhibit_y3
results %>% filter(p.value<0.05) %>% filter(outcome == "ch_briefpinhibit_y3") %>% select(term)

model_reduit_briefpinhibit <- lm(
  ch_briefpinhibit_y3 ~
    ch_age_SRS_BRIEFP_Y3 + 
    po_w_kg_3cat + 
    po_he_3cat + 
    mo_dipl_2cat + 
    mo_age + 
    mo_bmi_bepr_3cat + 
    ch_sex + 
    mo_par_2cat +
    ch_bf_duration_till48w_4cat + 
    po_gd + 
    po_delmod +
    ch_food_intro_Y1_3cat + 
    mo_pets + 
    ch_antibio_Y1_2cat + 
    home_total_y3 + 
    mo_hadtotscore_grt3_imp + 
    mo_tob_gr_anyt_yn_n2 +
    ch_tabacco_passive_up_to_Y1 + 
    ch_care_main_12m_opt2_2c,
  data = bdd_final_imp_1)

model_complet_briefpinhibit <- lm(
  ch_briefpinhibit_y3 ~
    ch_feces_rel_p4_std_Y1_10 +
    ch_age_SRS_BRIEFP_Y3 + 
    po_w_kg_3cat + 
    po_he_3cat + 
    mo_dipl_2cat + 
    mo_age + 
    mo_bmi_bepr_3cat + 
    ch_sex + 
    mo_par_2cat +
    ch_bf_duration_till48w_4cat + 
    po_gd + 
    po_delmod +
    ch_food_intro_Y1_3cat + 
    mo_pets + 
    ch_antibio_Y1_2cat + 
    home_total_y3 + 
    mo_hadtotscore_grt3_imp + 
    mo_tob_gr_anyt_yn_n2 +
    ch_tabacco_passive_up_to_Y1 + 
    ch_care_main_12m_opt2_2c,
  data = bdd_final_imp_1)

## ch_briefpshift_y3 
results %>% filter(p.value<0.05) %>% filter(outcome == "ch_briefpshift_y3") %>% select(term)

model_reduit_briefpshift <- lm(
  ch_briefpshift_y3 ~
    ch_age_SRS_BRIEFP_Y3 + 
    po_w_kg_3cat + 
    po_he_3cat + 
    mo_dipl_2cat + 
    mo_age + 
    mo_bmi_bepr_3cat + 
    ch_sex + 
    mo_par_2cat +
    ch_bf_duration_till48w_4cat + 
    po_gd + 
    po_delmod +
    ch_food_intro_Y1_3cat + 
    mo_pets + 
    ch_antibio_Y1_2cat + 
    home_total_y3 + 
    mo_hadtotscore_grt3_imp + 
    mo_tob_gr_anyt_yn_n2 +
    ch_tabacco_passive_up_to_Y1 + 
    ch_care_main_12m_opt2_2c,
  data = bdd_final_imp_1)

model_complet_briefpshift <- lm(
  ch_briefpshift_y3 ~
    ch_feces_rel_g47_imp_log_std_Y1 +
    ch_age_SRS_BRIEFP_Y3 + 
    po_w_kg_3cat + 
    po_he_3cat + 
    mo_dipl_2cat + 
    mo_age + 
    mo_bmi_bepr_3cat + 
    ch_sex + 
    mo_par_2cat +
    ch_bf_duration_till48w_4cat + 
    po_gd + 
    po_delmod +
    ch_food_intro_Y1_3cat + 
    mo_pets + 
    ch_antibio_Y1_2cat + 
    home_total_y3 + 
    mo_hadtotscore_grt3_imp + 
    mo_tob_gr_anyt_yn_n2 +
    ch_tabacco_passive_up_to_Y1 + 
    ch_care_main_12m_opt2_2c,
  data = bdd_final_imp_1)

## ch_briefpemocontrol_y3 
results %>% filter(p.value<0.05) %>% filter(outcome == "ch_briefpemocontrol_y3") %>% select(term)

model_reduit_briefpemocontrol <- lm(
  ch_briefpemocontrol_y3 ~
    ch_age_SRS_BRIEFP_Y3 + 
    po_w_kg_3cat + 
    po_he_3cat + 
    mo_dipl_2cat + 
    mo_age + 
    mo_bmi_bepr_3cat + 
    ch_sex + 
    mo_par_2cat +
    ch_bf_duration_till48w_4cat + 
    po_gd + 
    po_delmod +
    ch_food_intro_Y1_3cat + 
    mo_pets + 
    ch_antibio_Y1_2cat + 
    home_total_y3 + 
    mo_hadtotscore_grt3_imp + 
    mo_tob_gr_anyt_yn_n2 +
    ch_tabacco_passive_up_to_Y1 + 
    ch_care_main_12m_opt2_2c,
  data = bdd_final_imp_1)

model_complet_briefpemocontrol <- lm(
  ch_briefpemocontrol_y3 ~
    ch_feces_rel_p4_std_Y1_10 +    
    ch_feces_rel_g2_imp_log_std_Y1 + 
    ch_feces_rel_g8_imp_log_std_Y1 +
    ch_feces_rel_g36_imp_log_std_Y1 +
    ch_feces_rel_g43_imp_log_std_Y1 +
    ch_feces_rel_g45_imp_log_std_Y1 +
    ch_feces_rel_g47_imp_log_std_Y1 +
    ch_age_SRS_BRIEFP_Y3 + 
    po_w_kg_3cat + 
    po_he_3cat + 
    mo_dipl_2cat + 
    mo_age + 
    mo_bmi_bepr_3cat + 
    ch_sex + 
    mo_par_2cat +
    ch_bf_duration_till48w_4cat + 
    po_gd + 
    po_delmod +
    ch_food_intro_Y1_3cat + 
    mo_pets + 
    ch_antibio_Y1_2cat + 
    home_total_y3 + 
    mo_hadtotscore_grt3_imp + 
    mo_tob_gr_anyt_yn_n2 +
    ch_tabacco_passive_up_to_Y1 + 
    ch_care_main_12m_opt2_2c,
  data = bdd_final_imp_1)


## ch_briefpworkmemo_y3
results %>% filter(p.value<0.05) %>% filter(outcome == "ch_briefpworkmemo_y3") %>% select(term)

model_reduit_briefpworkmemo <- lm(
  ch_briefpworkmemo_y3 ~
    ch_age_SRS_BRIEFP_Y3 + 
    po_w_kg_3cat + 
    po_he_3cat + 
    mo_dipl_2cat + 
    mo_age + 
    mo_bmi_bepr_3cat + 
    ch_sex + 
    mo_par_2cat +
    ch_bf_duration_till48w_4cat + 
    po_gd + 
    po_delmod +
    ch_food_intro_Y1_3cat + 
    mo_pets + 
    ch_antibio_Y1_2cat + 
    home_total_y3 + 
    mo_hadtotscore_grt3_imp + 
    mo_tob_gr_anyt_yn_n2 +
    ch_tabacco_passive_up_to_Y1 + 
    ch_care_main_12m_opt2_2c,
  data = bdd_final_imp_1)

model_complet_briefpworkmemo <- lm(
  ch_briefpworkmemo_y3 ~
    ch_feces_rel_p3_std_Y1_10 +     
    ch_feces_rel_g1_imp_log_std_Y1 + 
    ch_feces_rel_g47_imp_log_std_Y1 +
    ch_feces_rel_g90_imp_log_std_Y1 +
    ch_age_SRS_BRIEFP_Y3 + 
    po_w_kg_3cat + 
    po_he_3cat + 
    mo_dipl_2cat + 
    mo_age + 
    mo_bmi_bepr_3cat + 
    ch_sex + 
    mo_par_2cat +
    ch_bf_duration_till48w_4cat + 
    po_gd + 
    po_delmod +
    ch_food_intro_Y1_3cat + 
    mo_pets + 
    ch_antibio_Y1_2cat + 
    home_total_y3 + 
    mo_hadtotscore_grt3_imp + 
    mo_tob_gr_anyt_yn_n2 +
    ch_tabacco_passive_up_to_Y1 + 
    ch_care_main_12m_opt2_2c,
  data = bdd_final_imp_1)

## ch_briefpplan_y3
results %>% filter(p.value<0.05) %>% filter(outcome == "ch_briefpplan_y3") %>% select(term)

model_reduit_briefpplan <- lm(
  ch_briefpplan_y3 ~
    ch_age_SRS_BRIEFP_Y3 + 
    po_w_kg_3cat + 
    po_he_3cat + 
    mo_dipl_2cat + 
    mo_age + 
    mo_bmi_bepr_3cat + 
    ch_sex + 
    mo_par_2cat +
    ch_bf_duration_till48w_4cat + 
    po_gd + 
    po_delmod +
    ch_food_intro_Y1_3cat + 
    mo_pets + 
    ch_antibio_Y1_2cat + 
    home_total_y3 + 
    mo_hadtotscore_grt3_imp + 
    mo_tob_gr_anyt_yn_n2 +
    ch_tabacco_passive_up_to_Y1 + 
    ch_care_main_12m_opt2_2c,
  data = bdd_final_imp_1)

model_complet_briefpplan <- lm(
  ch_briefpplan_y3 ~
    ch_feces_rel_p1_std_Y1_10 +      
    ch_feces_rel_p3_std_Y1_10 +     
    ch_feces_rel_g29_imp_log_std_Y1 +
    ch_age_SRS_BRIEFP_Y3 + 
    po_w_kg_3cat + 
    po_he_3cat + 
    mo_dipl_2cat + 
    mo_age + 
    mo_bmi_bepr_3cat + 
    ch_sex + 
    mo_par_2cat +
    ch_bf_duration_till48w_4cat + 
    po_gd + 
    po_delmod +
    ch_food_intro_Y1_3cat + 
    mo_pets + 
    ch_antibio_Y1_2cat + 
    home_total_y3 + 
    mo_hadtotscore_grt3_imp + 
    mo_tob_gr_anyt_yn_n2 +
    ch_tabacco_passive_up_to_Y1 + 
    ch_care_main_12m_opt2_2c,
  data = bdd_final_imp_1)

## ch_WPPSI_verbal_comprehension_cor_Y3 
results %>% filter(p.value<0.05) %>% filter(outcome == "ch_WPPSI_verbal_comprehension_cor_Y3") %>% select(term)

model_reduit_WPPSI_verbal_comprehension <- lm(
  ch_WPPSI_verbal_comprehension_cor_Y3 ~
    po_w_kg_3cat + 
    po_he_3cat + 
    mo_dipl_2cat + 
    mo_age + 
    mo_bmi_bepr_3cat + 
    ch_sex + 
    mo_par_2cat +
    ch_bf_duration_till48w_4cat + 
    po_gd + 
    po_delmod +
    ch_food_intro_Y1_3cat + 
    mo_pets + 
    ch_antibio_Y1_2cat + 
    home_total_y3 + 
    mo_hadtotscore_grt3_imp + 
    mo_tob_gr_anyt_yn_n2 +
    ch_tabacco_passive_up_to_Y1 + 
    ch_care_main_12m_opt2_2c,
  data = bdd_final_imp_1)

model_complet_WPPSI_verbal_comprehension <- lm(
  ch_WPPSI_verbal_comprehension_cor_Y3 ~
    ch_feces_rel_g11_imp_log_std_Y1 +
    ch_feces_rel_g20_imp_log_std_Y1 +
    po_w_kg_3cat + 
    po_he_3cat + 
    mo_dipl_2cat + 
    mo_age + 
    mo_bmi_bepr_3cat + 
    ch_sex + 
    mo_par_2cat +
    ch_bf_duration_till48w_4cat + 
    po_gd + 
    po_delmod +
    ch_food_intro_Y1_3cat + 
    mo_pets + 
    ch_antibio_Y1_2cat + 
    home_total_y3 + 
    mo_hadtotscore_grt3_imp + 
    mo_tob_gr_anyt_yn_n2 +
    ch_tabacco_passive_up_to_Y1 + 
    ch_care_main_12m_opt2_2c,
  data = bdd_final_imp_1)

## ch_WPPSI_visuospatiale_cor_Y3 
results %>% filter(p.value<0.05) %>% filter(outcome == "ch_WPPSI_visuospatiale_cor_Y3") %>% select(term)

## ch_WPPSI_work_memory_cor_Y3 
results %>% filter(p.value<0.05) %>% filter(outcome == "ch_WPPSI_work_memory_cor_Y3") %>% select(term)

model_reduit_WPPSI_work_memory <- lm(
  ch_WPPSI_work_memory_cor_Y3 ~
    po_w_kg_3cat + 
    po_he_3cat + 
    mo_dipl_2cat + 
    mo_age + 
    mo_bmi_bepr_3cat + 
    ch_sex + 
    mo_par_2cat +
    ch_bf_duration_till48w_4cat + 
    po_gd + 
    po_delmod +
    ch_food_intro_Y1_3cat + 
    mo_pets + 
    ch_antibio_Y1_2cat + 
    home_total_y3 + 
    mo_hadtotscore_grt3_imp + 
    mo_tob_gr_anyt_yn_n2 +
    ch_tabacco_passive_up_to_Y1 + 
    ch_care_main_12m_opt2_2c,
  data = bdd_final_imp_1)

model_complet_WPPSI_work_memory <- lm(
  ch_WPPSI_work_memory_cor_Y3 ~
    ch_feces_rel_g27_imp_log_std_Y1 +
    ch_feces_rel_g45_imp_log_std_Y1 +
    po_w_kg_3cat + 
    po_he_3cat + 
    mo_dipl_2cat + 
    mo_age + 
    mo_bmi_bepr_3cat + 
    ch_sex + 
    mo_par_2cat +
    ch_bf_duration_till48w_4cat + 
    po_gd + 
    po_delmod +
    ch_food_intro_Y1_3cat + 
    mo_pets + 
    ch_antibio_Y1_2cat + 
    home_total_y3 + 
    mo_hadtotscore_grt3_imp + 
    mo_tob_gr_anyt_yn_n2 +
    ch_tabacco_passive_up_to_Y1 + 
    ch_care_main_12m_opt2_2c,
  data = bdd_final_imp_1)

## ch_WPPSI_total_cor_Y3 
results %>% filter(p.value<0.05) %>% filter(outcome == "ch_WPPSI_total_cor_Y3") %>% select(term)

model_reduit_WPPSI_total <- lm(
  ch_WPPSI_total_cor_Y3 ~
    po_w_kg_3cat + 
    po_he_3cat + 
    mo_dipl_2cat + 
    mo_age + 
    mo_bmi_bepr_3cat + 
    ch_sex + 
    mo_par_2cat +
    ch_bf_duration_till48w_4cat + 
    po_gd + 
    po_delmod +
    ch_food_intro_Y1_3cat + 
    mo_pets + 
    ch_antibio_Y1_2cat + 
    home_total_y3 + 
    mo_hadtotscore_grt3_imp + 
    mo_tob_gr_anyt_yn_n2 +
    ch_tabacco_passive_up_to_Y1 + 
    ch_care_main_12m_opt2_2c,
  data = bdd_final_imp_1)

model_complet_WPPSI_total <- lm(
  ch_WPPSI_total_cor_Y3 ~
    ch_feces_rel_g20_imp_log_std_Y1 +
    po_w_kg_3cat + 
    po_he_3cat + 
    mo_dipl_2cat + 
    mo_age + 
    mo_bmi_bepr_3cat + 
    ch_sex + 
    mo_par_2cat +
    ch_bf_duration_till48w_4cat + 
    po_gd + 
    po_delmod +
    ch_food_intro_Y1_3cat + 
    mo_pets + 
    ch_antibio_Y1_2cat + 
    home_total_y3 + 
    mo_hadtotscore_grt3_imp + 
    mo_tob_gr_anyt_yn_n2 +
    ch_tabacco_passive_up_to_Y1 + 
    ch_care_main_12m_opt2_2c,
  data = bdd_final_imp_1)

### Assemblage ----
R2_complet_cbcl_int <- summary(model_complet_cbcl_int)$r.squared
R2_reduit_cbcl_int <- summary(model_reduit_cbcl_int)$r.squared

R2_complet_cbcl_ext <- summary(model_complet_cbcl_ext)$r.squared
R2_reduit_cbcl_ext <- summary(model_reduit_cbcl_ext)$r.squared

R2_complet_srs <- summary(model_complet_srs)$r.squared
R2_reduit_srs <- summary(model_reduit_srs)$r.squared

R2_complet_briefpinhibit <- summary(model_complet_briefpinhibit)$r.squared
R2_reduit_briefpinhibit <- summary(model_reduit_briefpinhibit)$r.squared

R2_complet_briefpshift <- summary(model_complet_briefpshift)$r.squared
R2_reduit_briefpshift <- summary(model_reduit_briefpshift)$r.squared

R2_complet_briefpemocontrol <- summary(model_complet_briefpemocontrol)$r.squared
R2_reduit_briefpemocontrol <- summary(model_reduit_briefpemocontrol)$r.squared

R2_complet_briefpworkmemo <- summary(model_complet_briefpworkmemo)$r.squared
R2_reduit_briefpworkmemo <- summary(model_reduit_briefpworkmemo)$r.squared

R2_complet_briefpplan <- summary(model_complet_briefpplan)$r.squared
R2_reduit_briefpplan <- summary(model_reduit_briefpplan)$r.squared

R2_complet_WPPSI_verbal_comprehension <- summary(model_complet_WPPSI_verbal_comprehension)$r.squared
R2_reduit_WPPSI_verbal_comprehension <- summary(model_reduit_WPPSI_verbal_comprehension)$r.squared

R2_complet_WPPSI_work_memory <- summary(model_complet_WPPSI_work_memory)$r.squared
R2_reduit_WPPSI_work_memory <- summary(model_reduit_WPPSI_work_memory)$r.squared

R2_complet_WPPSI_total <- summary(model_complet_WPPSI_total)$r.squared
R2_reduit_WPPSI_total <- summary(model_reduit_WPPSI_total)$r.squared

table_f2_Cohen <-
  list(
    ch_cbclintscore_y2 = (R2_complet_cbcl_int - R2_reduit_cbcl_int) / (1 - R2_complet_cbcl_int),
    ch_cbclextscore_y2 = (R2_complet_cbcl_ext - R2_reduit_cbcl_ext) / (1 - R2_complet_cbcl_ext),
    ch_SRStotal_y3 = (R2_complet_srs - R2_reduit_srs) / (1 - R2_complet_srs),
    ch_briefpinhibit_y3 = (R2_complet_briefpinhibit - R2_reduit_briefpinhibit) / (1 - R2_complet_briefpinhibit),
    ch_briefpshift_y3 = (R2_complet_briefpshift - R2_reduit_briefpshift) / (1 - R2_complet_briefpshift),
    ch_briefpemocontrol_y3 = (R2_complet_briefpemocontrol - R2_reduit_briefpemocontrol) / (1 - R2_complet_briefpemocontrol),
    ch_briefpworkmemo_y3 = (R2_complet_briefpworkmemo - R2_reduit_briefpworkmemo) / (1 - R2_complet_briefpworkmemo),
    ch_briefpplan_y3 = (R2_complet_briefpplan - R2_reduit_briefpplan) / (1 - R2_complet_briefpplan),
    ch_WPPSI_verbal_comprehension_cor_Y3 = (R2_complet_WPPSI_verbal_comprehension - R2_reduit_WPPSI_verbal_comprehension) / (1 - R2_complet_WPPSI_verbal_comprehension),
    ch_WPPSI_work_memory_cor_Y3 = (R2_complet_WPPSI_work_memory - R2_reduit_WPPSI_work_memory) / (1 - R2_complet_WPPSI_work_memory),
    ch_WPPSI_total_cor_Y3 = (R2_complet_WPPSI_total - R2_reduit_WPPSI_total) / (1 - R2_complet_WPPSI_total)) %>%
  unlist() %>%
  as.data.frame() %>%
  rownames_to_column(var = "outcome") %>%
  rename("f2_Cohen" = ".") %>%
  mutate(f2_Cohen = format(round(f2_Cohen, 2), digits = 2))

table_5 <- list(tidy(model_complet_cbcl_int, conf.int = TRUE) %>% mutate(outcome = "ch_cbclintscore_y2"), 
                tidy(model_complet_cbcl_ext, conf.int = TRUE) %>% mutate(outcome = "ch_cbclextscore_y2"), 
                tidy(model_complet_srs, conf.int = TRUE) %>% mutate(outcome = "ch_SRStotal_y3"), 
                tidy(model_complet_briefpinhibit, conf.int = TRUE) %>% mutate(outcome = "ch_briefpinhibit_y3"),  
                tidy(model_complet_briefpshift, conf.int = TRUE) %>% mutate(outcome = "ch_briefpshift_y3"), 
                tidy(model_complet_briefpemocontrol, conf.int = TRUE) %>% mutate(outcome = "ch_briefpemocontrol_y3"), 
                tidy(model_complet_briefpworkmemo, conf.int = TRUE) %>% mutate(outcome = "ch_briefpworkmemo_y3"), 
                tidy(model_complet_briefpplan, conf.int = TRUE) %>% mutate(outcome = "ch_briefpplan_y3"), 
                tidy(model_complet_WPPSI_verbal_comprehension, conf.int = TRUE) %>% mutate(outcome = "ch_WPPSI_verbal_comprehension_cor_Y3"), 
                tidy(model_complet_WPPSI_work_memory, conf.int = TRUE) %>% mutate(outcome = "ch_WPPSI_work_memory_cor_Y3"), 
                tidy(model_complet_WPPSI_total, conf.int = TRUE) %>% mutate(outcome = "ch_WPPSI_total_cor_Y3")) %>%
  bind_rows() %>%
  filter(term %in% c("ch_feces_rel_g19_imp_log_std_Y1",   # ch_cbclintscore_y2
                     "ch_feces_rel_g21_imp_log_std_Y1", 
                     "ch_feces_rel_g36_imp_log_std_Y1", 
                     "ch_feces_rel_g45_imp_log_std_Y1", 
                     
                     "ch_feces_rel_p4_std_Y1_10",        # ch_cbclextscore_y2
                     "ch_feces_rel_g1_imp_log_std_Y1", 
                     "ch_feces_rel_g13_imp_log_std_Y1",
                     "ch_feces_rel_g15_imp_log_std_Y1",
                     "ch_feces_rel_g19_imp_log_std_Y1",
                     "ch_feces_rel_g34_imp_log_std_Y1",
                     "ch_feces_rel_g39_imp_log_std_Y1",
                     "ch_feces_rel_g43_imp_log_std_Y1",
                     
                     "ch_feces_rel_p4_std_Y1_10",         # ch_SRStotal_y3
                     
                     "ch_feces_rel_p4_std_Y1_10",         # ch_briefpinhibit_y3
                     
                     "ch_feces_rel_g47_imp_log_std_Y1",   # ch_briefpshift_y3
                     
                     "ch_feces_rel_p4_std_Y1_10",        # ch_briefpemocontrol_y3
                     "ch_feces_rel_g2_imp_log_std_Y1",
                     "ch_feces_rel_g8_imp_log_std_Y1",
                     "ch_feces_rel_g36_imp_log_std_Y1", 
                     "ch_feces_rel_g43_imp_log_std_Y1", 
                     "ch_feces_rel_g45_imp_log_std_Y1",
                     "ch_feces_rel_g47_imp_log_std_Y1",
                     
                     "ch_feces_rel_p3_std_Y1_10",        # ch_briefpworkmemo_y3
                     "ch_feces_rel_g1_imp_log_std_Y1", 
                     "ch_feces_rel_g47_imp_log_std_Y1", 
                     "ch_feces_rel_g90_imp_log_std_Y1", 
                     
                     "ch_feces_rel_p1_std_Y1_10",       # ch_briefpplan_y3
                     "ch_feces_rel_p3_std_Y1_10",      
                     "ch_feces_rel_g29_imp_log_std_Y1", 
                     
                     "ch_feces_rel_g11_imp_log_std_Y1",   # ch_WPPSI_verbal_comprehension_cor_Y3
                     "ch_feces_rel_g20_imp_log_std_Y1", 
                     
                     "ch_feces_rel_g27_imp_log_std_Y1",   # ch_WPPSI_work_memory_cor_Y3
                     "ch_feces_rel_g45_imp_log_std_Y1", 
                     
                     "ch_feces_rel_g20_imp_log_std_Y1")) %>%  # ch_WPPSI_total_cor_Y3
  mutate(
    estimate = format(round(estimate, 2), digits = 2),
    conf.low = format(round(conf.low, 1), digits = 1),
    conf.high = format(round(conf.high, 1), digits = 1),
    "95% CI" = paste(conf.low, conf.high, sep = ","), 
    p.value = case_when(p.value < 0.001 ~ format(round(p.value, 4), nsmall = 4),
                        p.value < 0.01 ~ format(round(p.value, 3), nsmall = 3),
                        p.value > 0.01 ~ format(round(p.value, 2), nsmall = 2))) %>%
  select(outcome, explanatory = term, Beta = estimate, "95% CI", "p.value")

table_5 <- left_join(table_5, table_f2_Cohen, by = "outcome")

corres <- data.frame(Nom = genera_var_labels, Explanatory = genera_var_names, stringsAsFactors = FALSE)

table_5 <- left_join(table_5, corres, by = "explanatory")
table_5 <- table_5 %>%
  mutate(
    Nom = case_when(explanatory == "ch_feces_rel_p1_std_Y1_10" ~ "Firmicutes", 
                    explanatory == "ch_feces_rel_p3_std_Y1_10" ~ "Bacteroidetes", 
                    explanatory == "ch_feces_rel_p4_std_Y1_10" ~ "Proteobacteria", 
                    .default = Nom)) %>%
  select(outcome, explanatory, Nom, Beta, "95% CI", "p.value", f2_Cohen)

write.xlsx(table_5, file = "4_output/table_5.xlsx")


rm(model_complet_cbcl_int, model_reduit_cbcl_int, R2_complet_cbcl_int, R2_reduit_cbcl_int,
   model_complet_cbcl_ext, model_reduit_cbcl_ext, R2_complet_cbcl_ext, R2_reduit_cbcl_ext,
   model_complet_srs, model_reduit_srs, R2_complet_srs, R2_reduit_srs,
   model_complet_briefpinhibit, model_reduit_briefpinhibit, R2_complet_briefpinhibit, R2_reduit_briefpinhibit,
   model_complet_briefpshift, model_reduit_briefpshift, R2_complet_briefpshift, R2_reduit_briefpshift,
   model_complet_briefpemocontrol, model_reduit_briefpemocontrol, R2_complet_briefpemocontrol, R2_reduit_briefpemocontrol,
   model_complet_briefpworkmemo, model_reduit_briefpworkmemo, R2_complet_briefpworkmemo, R2_reduit_briefpworkmemo,
   model_complet_briefpplan, model_reduit_briefpplan, R2_complet_briefpplan, R2_reduit_briefpplan,
   model_complet_WPPSI_verbal_comprehension, model_reduit_WPPSI_verbal_comprehension, R2_complet_WPPSI_verbal_comprehension, R2_reduit_WPPSI_verbal_comprehension,
   model_complet_WPPSI_work_memory, model_reduit_WPPSI_work_memory, R2_complet_WPPSI_work_memory, R2_reduit_WPPSI_work_memory,
   model_complet_WPPSI_total, model_reduit_WPPSI_total, R2_complet_WPPSI_total, R2_reduit_WPPSI_total)

## Table 5 bis ----
corres <- results %>% 
  filter(p.value<0.05) %>% 
  select(term, term_2) %>% 
  unique() %>%
  mutate(
    term = factor(term,
                  levels = c(
                    "ch_feces_rel_p1_std_Y1_10", "ch_feces_rel_p3_std_Y1_10", "ch_feces_rel_p4_std_Y1_10",
                    "ch_feces_rel_g1_imp_log_std_Y1", "ch_feces_rel_g2_imp_log_std_Y1",
                    "ch_feces_rel_g8_imp_log_std_Y1", "ch_feces_rel_g11_imp_log_std_Y1",
                    "ch_feces_rel_g13_imp_log_std_Y1", "ch_feces_rel_g15_imp_log_std_Y1",
                    "ch_feces_rel_g19_imp_log_std_Y1", "ch_feces_rel_g20_imp_log_std_Y1",
                    "ch_feces_rel_g21_imp_log_std_Y1", "ch_feces_rel_g27_imp_log_std_Y1",
                    "ch_feces_rel_g29_imp_log_std_Y1", "ch_feces_rel_g34_imp_log_std_Y1",
                    "ch_feces_rel_g36_imp_log_std_Y1", "ch_feces_rel_g39_imp_log_std_Y1",
                    "ch_feces_rel_g43_imp_log_std_Y1", "ch_feces_rel_g45_imp_log_std_Y1",
                    "ch_feces_rel_g47_imp_log_std_Y1", "ch_feces_rel_g90_imp_log_std_Y1"))) %>%
  arrange(term)

## Visu des corrélations 
figure_SX <- bdd_final_imp_1 %>% 
  select("ch_feces_rel_p1_std_Y1_10", "ch_feces_rel_p3_std_Y1_10", "ch_feces_rel_p4_std_Y1_10",
         "ch_feces_rel_g1_imp_log_std_Y1", "ch_feces_rel_g2_imp_log_std_Y1",
         "ch_feces_rel_g8_imp_log_std_Y1", "ch_feces_rel_g11_imp_log_std_Y1",
         "ch_feces_rel_g13_imp_log_std_Y1", "ch_feces_rel_g15_imp_log_std_Y1",
         "ch_feces_rel_g19_imp_log_std_Y1", "ch_feces_rel_g20_imp_log_std_Y1",
         "ch_feces_rel_g21_imp_log_std_Y1", "ch_feces_rel_g27_imp_log_std_Y1",
         "ch_feces_rel_g29_imp_log_std_Y1", "ch_feces_rel_g34_imp_log_std_Y1",
         "ch_feces_rel_g36_imp_log_std_Y1", "ch_feces_rel_g39_imp_log_std_Y1",
         "ch_feces_rel_g43_imp_log_std_Y1", "ch_feces_rel_g45_imp_log_std_Y1",
         "ch_feces_rel_g47_imp_log_std_Y1", "ch_feces_rel_g90_imp_log_std_Y1") 

figure_SX <- cor(figure_SX, 
                 use = "pairwise.complete.obs", 
                 method = "pearson")

plot.new()
corrplot(figure_SX, 
         method = 'color', 
         type = "lower", 
         tl.col = 'black', 
         tl.srt = 45, 
         addCoef.col = "black",
         # number.cex = 0.5,
         # number.digits = 1,
         tl.cex = 0.5,
         col = rev(COL2(diverging = "RdYlBu")))
dev.off()

## Regressions linaires 
gm_interest <- 
  c("ch_feces_rel_p1_std_Y1_10", "ch_feces_rel_p3_std_Y1_10", "ch_feces_rel_p4_std_Y1_10",
    "ch_feces_rel_g1_imp_log_std_Y1", "ch_feces_rel_g2_imp_log_std_Y1",
    "ch_feces_rel_g8_imp_log_std_Y1", "ch_feces_rel_g11_imp_log_std_Y1",
    "ch_feces_rel_g13_imp_log_std_Y1", "ch_feces_rel_g15_imp_log_std_Y1",
    "ch_feces_rel_g19_imp_log_std_Y1", "ch_feces_rel_g20_imp_log_std_Y1",
    "ch_feces_rel_g21_imp_log_std_Y1", "ch_feces_rel_g27_imp_log_std_Y1",
    "ch_feces_rel_g29_imp_log_std_Y1", "ch_feces_rel_g34_imp_log_std_Y1",
    "ch_feces_rel_g36_imp_log_std_Y1", "ch_feces_rel_g39_imp_log_std_Y1",
    "ch_feces_rel_g43_imp_log_std_Y1", "ch_feces_rel_g45_imp_log_std_Y1",
    "ch_feces_rel_g47_imp_log_std_Y1", "ch_feces_rel_g90_imp_log_std_Y1")



Table_5_bis <- data.frame()

for (outcome in outcomes) {
  
  if (outcome %in% c("ch_cbclintscore_y2", "ch_cbclextscore_y2")) {
    
    formule <- as.formula(paste(outcome, "~", paste(c(gm_interest, "ch_age_CBCL_Y2", covariates), collapse = " + ")))
  } else if (outcome %in% c("ch_SRStotal_y3", "ch_briefpinhibit_y3", "ch_briefpshift_y3", 
                            "ch_briefpemocontrol_y3", "ch_briefpworkmemo_y3", 
                            "ch_briefpplan_y3")) {
    formule <- as.formula(paste(outcome, "~", paste(c(gm_interest, "ch_age_SRS_BRIEFP_Y3", covariates), collapse = " + ")))
  } else {
    formule <- as.formula(paste(outcome, "~", paste(c(gm_interest, covariates), collapse = " + ")))
  }
  
  model_full <- lm(formule, data = bdd_final_imp_1)  
  
  coeffs <- summary(model_full)$coefficients[gm_interest, ]
  
  beta <- coeffs[, "Estimate"]
  p_values <- coeffs[, "Pr(>|t|)"]
  confint <- confint(model_full, level = 0.95)[gm_interest, ]
  r_squared <- summary(model_full)$r.squared
  
  results <- data.frame(
    Outcome = outcome,
    Explanatory = rownames(coeffs),
    Beta = beta,
    conf.low = confint[, 1],
    conf.high = confint[, 2],
    p.value = p_values, 
    R2.full = r_squared  
  )
  
  Table_5_bis <- rbind(Table_5_bis, results) %>% as.data.frame()
}


# Dataframe pour stocker les résultats
Table_5_bis_reduced <- data.frame(
  Outcome = character(),
  Explanatory = character(),
  Beta = numeric(),
  conf.low = numeric(),
  conf.high = numeric(),
  p.value = numeric(),
  R2.red = numeric(),
  stringsAsFactors = FALSE
)

# Régression pour les outcomes spécifiques
for (outcome in outcomes) {
  if (outcome %in% c("ch_cbclintscore_y2", "ch_cbclextscore_y2")) {
    formula <- as.formula(paste(outcome, "~", paste(c("ch_age_CBCL_Y2", covariates), collapse = "+")))
  } else if (outcome %in% c("ch_SRStotal_y3", "ch_briefpinhibit_y3", "ch_briefpshift_y3", 
                            "ch_briefpemocontrol_y3", "ch_briefpworkmemo_y3", 
                            "ch_briefpplan_y3")) {
    formula <- as.formula(paste(outcome, "~", paste(c("ch_age_SRS_BRIEFP_Y3", covariates), collapse = "+")))
  } else {
    formula <- as.formula(paste(outcome, "~", paste(covariates, collapse = "+")))
  }
  
  # Ajustement du modèle
  model <- lm(formula, data = bdd_final_imp_1)
  
  # R² du modèle
  r_squared <- summary(model)$r.squared
  
  # Extraire les coefficients et p-values
  coefs <- summary(model)$coefficients
  for (var in rownames(coefs)) {
    beta <- coefs[var, "Estimate"]
    p_value <- coefs[var, "Pr(>|t|)"]
    ci <- confint(model)[var, ]
    
    # Ajouter les résultats au dataframe
    Table_5_bis_reduced <- rbind(Table_5_bis_reduced, data.frame(Outcome = outcome,
                                                                 Explanatory = var,
                                                                 Beta = beta,
                                                                 conf.low = ci[1],
                                                                 conf.high = ci[2],
                                                                 p.value = p_value,
                                                                 R2.red = r_squared,
                                                                 stringsAsFactors = FALSE))
  }
}

Table_5_bis <- Table_5_bis_reduced %>%
  select(Outcome, R2.red) %>%
  unique() %>%
  right_join(Table_5_bis, by = "Outcome")


Table_5_bis <- Table_5_bis %>%
  mutate(
    Beta = format(round(Beta, 2), digits = 2),
    conf.low = format(round(conf.low, 1), digits = 1),
    conf.high = format(round(conf.high, 1), digits = 1),
    "95% CI" = paste(conf.low, conf.high, sep = ","), 
    p.value = case_when(p.value < 0.001 ~ format(round(p.value, 4), nsmall = 4),
                        p.value < 0.01 ~ format(round(p.value, 3), nsmall = 3),
                        p.value > 0.01 ~ format(round(p.value, 2), nsmall = 2)), 
    Cohens_f2 = (R2.full - R2.red) / (1 - R2.full),
    R2.full = format(round(R2.full, 2), digits = 2),
    R2.red = format(round(R2.red, 2), digits = 2), 
    Cohens_f2 = format(round(Cohens_f2, 2), digits = 2)) %>%
  select(Outcome, Explanatory, Beta, "95% CI", "p.value", R2.full, R2.red, Cohens_f2)


Table_5_bis <- Table_5_bis %>%
  left_join(corres, by = "Explanatory") %>%
  mutate(
    Nom = case_when(Explanatory == "ch_feces_rel_p1_std_Y1_10" ~ "Firmicutes", 
                    Explanatory == "ch_feces_rel_p3_std_Y1_10" ~ "Bacteroidetes", 
                    Explanatory == "ch_feces_rel_p4_std_Y1_10" ~ "Proteobacteria", 
                    .default = Nom), 
    Nom = fct_recode(Nom,
                     "Clostridium IV" = "Clostridium_IV",
                     "Clostridium XlVa" = "Clostridium_XlVa",
                     "Clostridium XVIII" = "Clostridium_XVIII",
                     "Saccharibacteria genera incertae sedis" = "Saccharibacteria_genera_incertae_sedis")) %>%
  select(Outcome, Explanatory, Nom, everything())



write.xlsx(Table_5_bis, file = "4_output/table_5_bis.xlsx")




### Vérification des corrélations ----
## ch_cbclintscore_y2
figure_SX <- bdd_final_imp_1 %>% 
  select("ch_feces_rel_g19_imp_log_std_Y1", 
         "ch_feces_rel_g21_imp_log_std_Y1", 
         "ch_feces_rel_g36_imp_log_std_Y1", 
         "ch_feces_rel_g45_imp_log_std_Y1") 

figure_SX <- cor(figure_SX, 
                 use = "pairwise.complete.obs", 
                 method = "pearson")

plot.new()
corrplot(figure_SX, 
         method = 'color', 
         type = "lower", 
         tl.col = 'black', 
         tl.srt = 45, 
         addCoef.col = "black",
         # number.cex = 0.5,
         # number.digits = 1,
         tl.cex = 0.5,
         col = rev(COL2(diverging = "RdYlBu")))
dev.off()

## ch_cbclextscore_y2
figure_SX <- bdd_final_imp_1 %>% 
  select("ch_feces_rel_p4_std_Y1_10",      
         "ch_feces_rel_g1_imp_log_std_Y1", 
         "ch_feces_rel_g13_imp_log_std_Y1",
         "ch_feces_rel_g15_imp_log_std_Y1",
         "ch_feces_rel_g19_imp_log_std_Y1",
         "ch_feces_rel_g34_imp_log_std_Y1",
         "ch_feces_rel_g39_imp_log_std_Y1",
         "ch_feces_rel_g43_imp_log_std_Y1") 

figure_SX <- cor(figure_SX, 
                 use = "pairwise.complete.obs", 
                 method = "pearson")

plot.new()
corrplot(figure_SX, 
         method = 'color', 
         type = "lower", 
         tl.col = 'black', 
         tl.srt = 45, 
         addCoef.col = "black",
         # number.cex = 0.5,
         # number.digits = 1,
         tl.cex = 0.5,
         col = rev(COL2(diverging = "RdYlBu")))
dev.off()

## ch_briefpemocontrol_y3
figure_SX <- bdd_final_imp_1 %>% 
  select("ch_feces_rel_p4_std_Y1_10",        
         "ch_feces_rel_g2_imp_log_std_Y1",
         "ch_feces_rel_g8_imp_log_std_Y1",
         "ch_feces_rel_g36_imp_log_std_Y1", 
         "ch_feces_rel_g43_imp_log_std_Y1", 
         "ch_feces_rel_g45_imp_log_std_Y1",
         "ch_feces_rel_g47_imp_log_std_Y1") 

figure_SX <- cor(figure_SX, 
                 use = "pairwise.complete.obs", 
                 method = "pearson")

plot.new()
corrplot(figure_SX, 
         method = 'color', 
         type = "lower", 
         tl.col = 'black', 
         tl.srt = 45, 
         addCoef.col = "black",
         # number.cex = 0.5,
         # number.digits = 1,
         tl.cex = 0.5,
         col = rev(COL2(diverging = "RdYlBu")))
dev.off()


## ch_briefpworkmemo_y3
figure_SX <- bdd_final_imp_1 %>% 
  select("ch_feces_rel_p3_std_Y1_10",        
         "ch_feces_rel_g1_imp_log_std_Y1", 
         "ch_feces_rel_g47_imp_log_std_Y1", 
         "ch_feces_rel_g90_imp_log_std_Y1") 

figure_SX <- cor(figure_SX, 
                 use = "pairwise.complete.obs", 
                 method = "pearson")

plot.new()
corrplot(figure_SX, 
         method = 'color', 
         type = "lower", 
         tl.col = 'black', 
         tl.srt = 45, 
         addCoef.col = "black",
         # number.cex = 0.5,
         # number.digits = 1,
         tl.cex = 0.5,
         col = rev(COL2(diverging = "RdYlBu")))
dev.off()

## ch_briefpplan_y3
figure_SX <- bdd_final_imp_1 %>% 
  select("ch_feces_rel_p1_std_Y1_10",       
         "ch_feces_rel_p3_std_Y1_10",      
         "ch_feces_rel_g29_imp_log_std_Y1") 

figure_SX <- cor(figure_SX, 
                 use = "pairwise.complete.obs", 
                 method = "pearson")

plot.new()
corrplot(figure_SX, 
         method = 'color', 
         type = "lower", 
         tl.col = 'black', 
         tl.srt = 45, 
         addCoef.col = "black",
         # number.cex = 0.5,
         # number.digits = 1,
         tl.cex = 0.5,
         col = rev(COL2(diverging = "RdYlBu")))
dev.off()

## ch_WPPSI_verbal_comprehension_cor_Y3
figure_SX <- bdd_final_imp_1 %>% 
  select("ch_feces_rel_g11_imp_log_std_Y1",   
         "ch_feces_rel_g20_imp_log_std_Y1") 

figure_SX <- cor(figure_SX, 
                 use = "pairwise.complete.obs", 
                 method = "pearson")

plot.new()
corrplot(figure_SX, 
         method = 'color', 
         type = "lower", 
         tl.col = 'black', 
         tl.srt = 45, 
         addCoef.col = "black",
         # number.cex = 0.5,
         # number.digits = 1,
         tl.cex = 0.5,
         col = rev(COL2(diverging = "RdYlBu")))
dev.off()

## ch_WPPSI_work_memory_cor_Y3
figure_SX <- bdd_final_imp_1 %>% 
  select("ch_feces_rel_g27_imp_log_std_Y1",   
         "ch_feces_rel_g45_imp_log_std_Y1") 

figure_SX <- cor(figure_SX, 
                 use = "pairwise.complete.obs", 
                 method = "pearson")

plot.new()
corrplot(figure_SX, 
         method = 'color', 
         type = "lower", 
         tl.col = 'black', 
         tl.srt = 45, 
         addCoef.col = "black",
         # number.cex = 0.5,
         # number.digits = 1,
         tl.cex = 0.5,
         col = rev(COL2(diverging = "RdYlBu")))
dev.off()



# Figures ----
## Fig.1: Forestplot phyla ----
figure_1 <- table_multi %>% 
  filter(`Gut microbiota parameters` %in% c("Firmicutes",
                                            "Actinobacteria", 
                                            "Bacteroidetes", 
                                            "Proteobacteria")) %>%
  mutate(Beta = as.numeric(Beta), 
         Outcome_rec = gsub("WPPSI", "WPPSI-IV", Outcome_rec), 
         Outcome_rec = gsub("SRS", "SRS-II", Outcome_rec), 
         Outcome_rec =
           fct_relevel(Outcome_rec,
                       "Total WPPSI-IV Y3", "Work memory WPPSI-IV Y3",
                       "Visuospatial WPPSI-IV Y3","Verbal comprehension WPPSI-IV Y3",
                       "Plan and organization BRIEF-P Y3", "Work memory BRIEF-P Y3",
                       "Emotional control BRIEF-P Y3","Shift BRIEF-P Y3",
                       "Inhibition BRIEF-P Y3", "Total SRS-II Y3",
                       "Externalizing CBCL Y2", "Internalizing CBCL Y2"),
        # Outcome_rec = fct_rev(Outcome_rec),
         `Gut microbiota parameters` =
           fct_relevel(`Gut microbiota parameters`, 
                       "Firmicutes", "Actinobacteria", "Bacteroidetes", "Proteobacteria")) %>%
  ggplot(aes(x = Outcome_rec, 
             y = Beta, 
             min = lower_CI, 
             ymax = upper_CI, 
             color = p_value_shape)) +
  geom_hline(yintercept = 0, linetype="dashed") +
  geom_pointrange(
    position = position_dodge(width = 0.5), 
    size = 0.4) +
  labs(x = "Neurodevelopement", y = "", color = "p-value") +
  theme_lucid() +
  coord_flip()  +
  facet_wrap(`Gut microbiota parameters`~.,  ncol = 4) +
  theme(
    legend.position = "right",
    legend.box = "vertical", 
    legend.justification = "right")  +
  scale_color_manual(values = c("p-value<0.05" = "red", "p-value≥0.05" = "black"))
figure_1 <- ggdraw(figure_1) +
  draw_label("All the p-values\nwere above the\nmultiple testing-\ncorrected threshold", 
           x = 0.865, y = 0.4, hjust = 0, size = 10, fontfamily = "sans", color = "gray30")

figure_1
ggsave("4_output/figures/Fig.1 forest_plot_phyla.tiff", 
       figure_1, 
       device = "tiff",
       units = "cm",
       dpi = 300,
       height = 15, 
       width = 25)

## Fig.2: Mahatan plot genera ----
figure_2 <- table_multi %>%
  filter(!`Gut microbiota parameters` %in% c("Firmicutes",
                                             "Actinobacteria",
                                             "Bacteroidetes",
                                             "Proteobacteria",
                                             "Shannon diversity",
                                             "Specific richness")) %>%
  mutate(
    `Gut microbiota parameters` = 
      fct_relevel(`Gut microbiota parameters`, 
                  "Saccharibacteria genera incertae sedis", "Peptoniphilus",
                  "Granulicatella", "Anaerotruncus", "Lactococcus", "Terrisporobacter",
                  "Oscillibacter", "Haemophilus", "Coprococcus", "Erysipelotrichaceae incertae sedis",
                  "Butyricicoccus", "Dialister", "Subdoligranulum", "Intestinibacter",
                  "Klebsiella", "Eisenbergiella", "Hungatella", "Dorea", "Eggerthella",
                  "Romboutsia", "Clostridium IV", "Ruminococcus 2", "Flavonifractor",
                  "Alistipes", "Collinsella", "Parabacteroides", "Fusicatenibacter",
                  "Veillonella", "Roseburia", "Enterobacter", "Cellulosibacter",
                  "Enterococcus", "Clostridium sensu stricto", "Clostridium XVIII",
                  "Ruminococcus", "Gemmiger", "Anaerostipes", "Lachnospiracea incertae sedis",
                  "Clostridium XlVa", "Streptococcus", "Faecalibacterium", "Akkermansia",
                  "Escherichia and Shigella", "Blautia", "Bacteroides", "Bifidobacterium"), 
    Outcome_rec = gsub("WPPSI", "WPPSI-IV", Outcome_rec), 
    Outcome_rec = gsub("SRS", "SRS-II", Outcome_rec)) %>%
  ggplot(aes(x = -log10(`p-value`), y = `Gut microbiota parameters`)) +
  geom_point(aes(shape = improved_neuro), size = 2) +
  geom_vline(xintercept = -log10(0.05), linetype = "dashed", color = "red") +
  geom_vline(xintercept = -log10(0.05/(33*7)), linetype = "dashed", color = "blue") +
  theme_lucid() +
  labs(x = "-log10(P-value)", 
       y = "Genera", 
       shape = "") +
  geom_text_repel(aes(label = ifelse(`p-value` < 0.05, as.character(Outcome_rec), "")), 
                  #segment.color = NA,   # supprimes les traits qui relient les points aux textes 
                  direction = "y",      # argument pour faire les ajustements de position seulement dans le sens vertical
                  #angle = 20, 
                  seed = 42,            # seed pour obtenir toujours le même ajustement 
                  #box.padding = 0.5,
                  hjust = -0.05,        # place le texte un tout petit peu au dessus des points
                  vjust = 0.5, 
                  nudge_x = 0.005, 
                  min.segment.length = 0.5  # ne met que les traits de plus de 0.5
  ) +
  scale_shape_manual(values = c("Altered neurodevelopmental outcome" = 15, "Improved neurodevelopmental outcome" = 17)) +# 15: carré plein, 17: triangle plein
  scale_x_continuous(limits = c(-log10(1), -log10(0.0001))) +
  theme(
    legend.position = "bottom",
    legend.box = "vertical", 
    legend.justification = "center", 
    axis.text.y = element_text(face = "italic")) + 
  annotate("text", 
           x = -log10(0.05/(33*7)), y = 0,
           label = "Multiple testing-\ncorrected threshold", 
           color = "blue", angle = 0, size = 3.5, hjust = 0.5, vjust = 1.5) +
  coord_cartesian(ylim = c(0, 46), clip = "off")

figure_2
ggsave("4_output/figures/Fig.2 manhattan_plot_genera.tiff", 
       figure_2, 
       device = "tiff",
       units = "cm",
       dpi = 300,
       height = 25, 
       width = 40)

## Fig.3: Forestplot final genera ----
figure_3 <- table_multi %>% 
  mutate(Beta = as.numeric(Beta), 
         Outcome_rec = gsub("WPPSI", "WPPSI-IV", Outcome_rec), 
         Outcome_rec = gsub("SRS", "SRS-II", Outcome_rec), 
         Outcome_rec = 
           fct_relevel(Outcome_rec,
                       "Internalizing CBCL Y2", 
                       "Externalizing CBCL Y2",
                       "Total SRS-II Y3", 
                       "Inhibition BRIEF-P Y3", 
                       "Shift BRIEF-P Y3",
                       "Emotional control BRIEF-P Y3",
                       "Work memory BRIEF-P Y3",
                       "Plan and organization BRIEF-P Y3", 
                       "Verbal comprehension WPPSI-IV Y3", 
                       "Visuospatial WPPSI-IV Y3",
                       "Work memory WPPSI-IV Y3", 
                       "Total WPPSI-IV Y3"), 
         `Gut microbiota parameters` = 
           fct_relevel(`Gut microbiota parameters`, 
                       "Saccharibacteria genera incertae sedis", "Peptoniphilus",
                       "Granulicatella", "Anaerotruncus", "Lactococcus", "Terrisporobacter",
                       "Oscillibacter", "Haemophilus", "Erysipelotrichaceae incertae sedis",
                       "Butyricicoccus", "Dialister", "Subdoligranulum", "Intestinibacter",
                       "Klebsiella", "Eisenbergiella", "Hungatella", "Dorea", "Eggerthella",
                       "Romboutsia", "Clostridium IV", "Ruminococcus 2", "Flavonifractor",
                       "Alistipes", "Collinsella", "Parabacteroides", "Fusicatenibacter",
                       "Roseburia", "Enterobacter", "Cellulosibacter", "Enterococcus",
                       "Coprococcus", "Veillonella", "Clostridium sensu stricto", "Clostridium XVIII",
                       "Ruminococcus", "Gemmiger", "Anaerostipes", "Lachnospiracea incertae sedis",
                       "Clostridium XlVa", "Streptococcus", "Faecalibacterium", "Akkermansia",
                       "Escherichia and Shigella", "Blautia", "Bacteroides", "Bifidobacterium",
                       "Proteobacteria", "Bacteroidetes", "Actinobacteria", "Firmicutes",
                       "Shannon diversity", "Specific richness"), 
         Phyla_corres = fct_recode(Phyla_corres, "Candidatus Saccharibacteria" = "Candidatus_Saccharibacteria"), 
         Phyla_corres = fct_relevel(Phyla_corres,
                                    "Firmicutes", "Actinobacteria", "Bacteroidetes", "Proteobacteria",
                                    "Verrucomicrobia", "Candidatus Saccharibacteria"), 
         `Gut microbiota parameters` = 
           fct_recode(`Gut microbiota parameters`, 
                      "Saccharibacteria" = "Saccharibacteria genera incertae sedis")) %>%
  filter(`p-value`<0.05) %>% 
  filter(!`Gut microbiota parameters` %in% c("Firmicutes",
                                             "Actinobacteria",
                                             "Bacteroidetes",
                                             "Proteobacteria",
                                             "Shannon diversity",
                                             "Specific richness")) %>%
  ggplot(aes(x = `Gut microbiota parameters`, 
             y = Beta, 
             min = lower_CI, 
             ymax = upper_CI, 
             color = Phyla_corres, 
             label = ifelse(`p-value` < 0.05, as.character(`Gut microbiota parameters`), ""))) +
  geom_hline(yintercept = 0, linetype="dashed") +
  geom_pointrange(
    size = 0.4,
    position = position_dodge(width = 1.2, preserve = "total")) +
  geom_text_repel(
    fontface = "italic", 
    nudge_x  = 0.3,
    aes(label = ifelse(`p-value` < 0.05, as.character(`Gut microbiota parameters`), "")) 
    #         fontface = "italic", 
    #         hjust = 0.5, 
    #         vjust = -0.5, 
    #         angle = 0, 
    #         size = 4, 
    #         position = position_dodge(width = 1.2, preserve = "total")
  ) +
  coord_flip()  +
  facet_grid(.~Outcome_rec, 
             scales = "free_y", 
             space = "free_y", 
             switch = "y", 
             labeller = as_labeller(function(labels) {
               labels <- gsub("Emotional control BRIEF-P Y3", "Emotional control\nBRIEF-P Y3", labels)
               labels <- gsub("Work memory", "Work memory\n", labels)
               labels <- gsub("Plan and organization BRIEF-P Y3", "Plan and organization\nBRIEF-P Y3", labels)
               labels <- gsub("Verbal comprehension WPPSI-IV Y3", "Verbal comprehension\nWPPSI-IV Y3", labels)
               return(labels)
             })
  ) + 
  labs(x = "", y = "", color = "Corresponding phylum") +
  scale_color_manual(values = c(
    "Firmicutes" = "black", 
    "Actinobacteria" = "#1b1b8e", # dark blue
    "Bacteroidetes" = "#1b8e1b", # dark green
    "Proteobacteria" = "darkred", # dark red
    "Verrucomicrobia" = "#6b1b8e", # dark purple
    "Candidatus Saccharibacteria" = "#8e6b1b")) +   # dark brown
  theme_lucid()  + 
  theme(axis.text.y = element_blank(), 
        strip.text.y.left = element_text(angle = 0, hjust = 1, size = 10), 
        panel.background = element_rect(color = "gray", fill = NULL), 
        legend.position = "bottom", 
        legend.title = element_text(face = "bold"))
figure_3


ggsave("4_output/figures/Fig.3 forest_plot_genera.tiff", 
       figure_3, 
       device = "tiff",
       units = "cm",
       dpi = 300,
       height = 20, 
       width = 34)


# Additional tables ----
## Table S1: Distribution gut microbiota parameters ----
table_S1 <- descrip_num(data = bdd, 
                        vars = c("ch_feces_SpecRich_5000_ASV_Y1",
                                 "ch_feces_Shannon_5000_ASV_Y1",
                                 "ch_feces_rel_p1_Y1",         
                                 "ch_feces_rel_p2_Y1",
                                 "ch_feces_rel_p3_Y1",
                                 "ch_feces_rel_p4_Y1",
                                 genera_var_names_raw))
#write.xlsx(table_S1, file = "4_output/Table_S1.xlsx")

## Table S2: Distribution neurodevelopment ----
table_S2 <- descrip_num(data = bdd_final, vars = outcomes)
#write.xlsx(table_S2, file = "4_output/Table_S2.xlsx")

## Table S3: Effects of the covariates on the outcomes ----
model_covars <- function(var_outcome, var_age, covars, data) {
  
  effectif <- data %>%                   # Création d'une colonne effectif
    filter(!is.na({{var_outcome}}))%>%
    select(all_of({{covars}}), 
           {{var_age}}) %>%
    tbl_summary(
      missing = "no", 
      type = list(ch_antibio_Y1_2cat ~ "categorical", 
                  home_total_y3 ~ "continuous"), 
      statistic = all_continuous() ~ "{N_nonmiss}") %>%
    bold_labels()
  
  tbl_model_univ <- data %>%                      # Création des colonnes modèle simple 
    select(
      {{var_outcome}}, 
      {{var_age}},
      all_of({{covars}})) %>%
    
    tbl_uvregression(
      method = lm ,
      y = {{var_outcome}},
      formula = "{y} ~ {x}", 
      hide_n = TRUE, 
      estimate_fun = scales::label_number(accuracy = .01, decimal.mark = "."),
      pvalue_fun = custom_pvalue_fun) %>%
    add_global_p() %>%
    bold_labels() %>%
    bold_p(t = 0.1)
  
  variables_explicatives <- c(covars, var_age)
  formule_regr <- as.formula(paste(var_outcome, "~", paste(variables_explicatives, collapse = " + ")))
  modele <- lm(formule_regr, data = data)
  tbl_model_multi <-  tbl_regression(
    modele, 
    estimate_fun = scales::label_number(accuracy = .01, decimal.mark = "."),
    pvalue_fun = custom_pvalue_fun) %>%
    add_global_p() %>%
    bold_labels() %>%
    bold_p(t = 0.1)
  
  tbl_model <- tbl_merge(
    tbls = list(effectif, 
                tbl_model_univ, 
                tbl_model_multi), 
    tab_spanner = c("", "**Univariate models**", "**Multivariate model**"))%>%
    modify_table_body(~.x %>% arrange(row_type == "glance_statistic"))
  
  return(tbl_model)
}

table_S3 <- tbl_merge(
  tbls = list(
    model_covars(var_outcome = "ch_cbclintscore_y2", var_age = "ch_age_CBCL_Y2", covars = covariates, data = bdd_final_imp_1),
    model_covars(var_outcome = "ch_cbclextscore_y2", var_age = "ch_age_CBCL_Y2", covars = covariates, data = bdd_final_imp_1),
    model_covars(var_outcome = "ch_SRStotal_y3", var_age = "ch_age_SRS_BRIEFP_Y3", covars = covariates, data = bdd_final_imp_1),
    model_covars(var_outcome = "ch_briefpinhibit_y3", var_age = "ch_age_SRS_BRIEFP_Y3", covars = covariates, data = bdd_final_imp_1),
    model_covars(var_outcome = "ch_briefpshift_y3", var_age = "ch_age_SRS_BRIEFP_Y3", covars = covariates, data = bdd_final_imp_1),
    model_covars(var_outcome = "ch_briefpemocontrol_y3", var_age = "ch_age_SRS_BRIEFP_Y3", covars = covariates, data = bdd_final_imp_1),
    model_covars(var_outcome = "ch_briefpworkmemo_y3", var_age = "ch_age_SRS_BRIEFP_Y3", covars = covariates, data = bdd_final_imp_1),
    model_covars(var_outcome = "ch_briefpplan_y3", var_age = "ch_age_SRS_BRIEFP_Y3", covars = covariates, data = bdd_final_imp_1),
    model_covars(var_outcome = "ch_WPPSI_verbal_comprehension_cor_Y3", var_age = "ch_WPPSI_psy_Y3", covars = covariates, data = bdd_final_imp_1),
    model_covars(var_outcome = "ch_WPPSI_visuospatiale_cor_Y3", var_age = "ch_WPPSI_psy_Y3", covars = covariates, data = bdd_final_imp_1),
    model_covars(var_outcome = "ch_WPPSI_work_memory_cor_Y3", var_age = "ch_WPPSI_psy_Y3", covars = covariates, data = bdd_final_imp_1), 
    model_covars(var_outcome = "ch_WPPSI_total_cor_Y3", var_age = "ch_WPPSI_psy_Y3", covars = covariates, data = bdd_final_imp_1)),
  tab_spanner = spanner_names)


## Table S4: Associations genera and neurodev ----
# Adjusted associations between the 46 most abundant genera in the child gut microbiota at one year and the neurodevelopment (n between X and X).
table_S4 <- tbl_stack(
  tbls = lapply(7:52, function(j) {  # tbl 7 à 52 : correspond aux analyses de genres
    tbl_merge(
      tbls = lapply(tbls_by_outcome_multi, function(i) i[[j]]),
      tab_spanner = spanner_names)}))

## Table S5: Sensitivity analysis - effect of the rarefaction threshold ----
# Sensitivity analysis - Adjusted associations between the gut microbiota α-diversity at different sequencing depths and the neurodevelopment. 
alpha_vec <- c("ch_feces_SpecRich_5000_ASV_std_Y1_10", "ch_feces_SpecRich_10000_ASV_std_Y1_10", 
               "ch_feces_Shannon_5000_ASV_std_Y1", "ch_feces_Shannon_10000_ASV_std_Y1")
table_S5 <- vector("list", length(outcomes))
names(table_S5) <- outcomes

# Boucle principale
for (outcome in outcomes) {
  # Sélection des covariables appropriées
  if (outcome %in% c("ch_cbclintscore_y2", "ch_cbclextscore_y2")) {
    covars <- covariates_map$CBCL
  } else if (outcome %in% c("ch_SRStotal_y3", "ch_briefpinhibit_y3", "ch_briefpshift_y3", "ch_briefpemocontrol_y3", "ch_briefpworkmemo_y3", "ch_briefpplan_y3")) {
    covars <- covariates_map$SRS_BRIEF
  } else if (outcome %in% c("ch_WPPSI_verbal_comprehension_cor_Y3", "ch_WPPSI_visuospatiale_cor_Y3", "ch_WPPSI_work_memory_cor_Y3", "ch_WPPSI_total_cor_Y3")) {
    covars <- covariates_map$IQ
  }
  
  tbls_for_outcome_multi_sensi_seuil <- vector("list", length(alpha_vec))
  names(tbls_for_outcome_multi_sensi_seuil) <- alpha_vec
  
  for (exposure in alpha_vec) {                                               # running linear regression
    terms <- c(exposure, covars)
    formula <- reformulate(terms, response = outcome)
    model <- lm(formula, data = bdd_final_imp_1_sensi_seuil)
    # model <- with(data = bdd_final_imp, 
    #               exp = lm(formula))
    
    tbl <-                                                                      
      tbl_regression(
        model, 
        include = exposure,
        estimate_fun = scales::label_number(accuracy = .01, decimal.mark = "."),
        pvalue_fun = custom_pvalue_fun,
        exponentiate = FALSE) %>%
      bold_p() %>%
      bold_labels() %>%
      add_global_p(include = exposure, singular.ok = TRUE, keep = TRUE) %>%
      add_n()%>%
      modify_table_body(~.x %>% arrange(row_type == "glance_statistic"))
    
    tbls_for_outcome_multi_sensi_seuil[[exposure]] <- tbl
  }
  table_S5[[outcome]] <- tbls_for_outcome_multi_sensi_seuil
}

table_S5_large <- tbl_stack(
  tbls = list(
    tbl_merge(tbls = lapply(tbls_by_outcome_multi, function(x) x$ch_feces_SpecRich_5000_ASV_std_Y1_10), # analyse principale seuil 5000
              tab_spanner = spanner_names),  
    tbl_merge(tbls = lapply(table_S5, function(x) x$ch_feces_SpecRich_5000_ASV_std_Y1_10),              # analyse de sensibilité seuil 5000 pour n sur seuil 10000
              tab_spanner = spanner_names),               
    tbl_merge(tbls = lapply(table_S5, function(x) x$ch_feces_SpecRich_10000_ASV_std_Y1_10),             # analyse de sensibilité seuil 10000 pour n sur seuil 10000
              tab_spanner = spanner_names),               
    tbl_merge(tbls = lapply(tbls_by_outcome_multi, function(x) x$ch_feces_Shannon_5000_ASV_std_Y1),     # analyse principale
              tab_spanner = spanner_names),  
    tbl_merge(tbls = lapply(table_S5, function(x) x$ch_feces_Shannon_5000_ASV_std_Y1),                  # analyse de sensibilité seuil 5000 pour n sur seuil 10000
              tab_spanner = spanner_names),               
    tbl_merge(tbls = lapply(table_S5, function(x) x$ch_feces_Shannon_10000_ASV_std_Y1),                 # analyse de sensibilité seuil 10000 pour n sur seuil 10000
              tab_spanner = spanner_names)))              


merge_tbls_function_rich <- function(index, tbls_by_outcome_multi, table_S5) {
  tbl_merge(
    tbls = list(
      tbls_by_outcome_multi[[index]]$ch_feces_SpecRich_5000_ASV_std_Y1_10, 
      table_S5[[index]]$ch_feces_SpecRich_5000_ASV_std_Y1_10,             
      table_S5[[index]]$ch_feces_SpecRich_10000_ASV_std_Y1_10), 
    tab_spanner = c("**Threshold 5,000 (n=350)**","**Threshold 5,000 (n=339)**", "**Threshold 10,000 (n=339)**"))
}

merge_tbls_function_sha <- function(index, tbls_by_outcome_multi, table_S5) {
  tbl_merge(
    tbls = list(
      tbls_by_outcome_multi[[index]]$ch_feces_Shannon_5000_ASV_std_Y1, 
      table_S5[[index]]$ch_feces_Shannon_5000_ASV_std_Y1,            
      table_S5[[index]]$ch_feces_Shannon_10000_ASV_std_Y1), 
    tab_spanner = c("**Threshold 5,000 (n=350)**","**Threshold 5,000 (n=339)**", "**Threshold 10,000 (n=339)**"))
}

table_S5_long <- 
  tbl_stack(
    tbls = c(lapply(1:12, function(index) merge_tbls_function_rich(index, tbls_by_outcome_multi, table_S5)), 
             lapply(1:12, function(index) merge_tbls_function_sha(index, tbls_by_outcome_multi, table_S5))),
    group_header = c(spanner_names, spanner_names))

table_S5 <- list(table_S5_long = table_S5_long, 
                 table_S5_large = table_S5_large)

rm(terms, formula, model, exposure, outcome, alpha_vec, tbl, covars)
rm(merge_tbls_function_rich, merge_tbls_function_sha, table_S5_large, table_S5_long, tbls_for_outcome_multi)

## Table S6: Sensitivity analysis – effect of the HOME Y3 variable on CBCL Y2 -----
## Sensitivity analysis – Effects of the HOME covariate assessed at 3 years on the CBCL outcomes assessed at 2 years.
covariates_sensi_home <- 
  bdd_final_imp_1 %>%
  select(
    ch_age_CBCL_Y2,
    all_of(covariates)) %>%
  select(-home_total_y3) %>%
  colnames()
cbcl_vec <- c("ch_cbclintscore_y2", "ch_cbclextscore_y2")

# Initialisation de la liste pour stocker les résultats
prep_table_S6 <- list()

for (outcome in cbcl_vec) {
  # Initialisation de la sous-liste pour stocker les résultats pour chaque outcome
  table_S6_for_outcome <- list()
  
  for (explicative in explanatory) {
    # Construction de la formule
    formula <- as.formula(paste(outcome, "~", explicative, "+", paste(covariates_sensi_home, collapse = "+")))
    
    # Création du modèle de régression linéaire
    model <- lm(formula, data = bdd_final_imp_1)
    
    # Création du tbl_regression sans les covariates_sensi_home
    tbl <-                                                                      
      tbl_regression(
        model, 
        include = explicative,
        estimate_fun = scales::label_number(accuracy = .01, decimal.mark = "."),
        pvalue_fun = custom_pvalue_fun,
        exponentiate = FALSE) %>%
      bold_p() %>%
      bold_labels() %>%
      add_n()
    
    table_S6_for_outcome[[explicative]] <- tbl        # Stockage du tbl_regression dans la liste
  }
  
  prep_table_S6[[outcome]] <- table_S6_for_outcome    # Stockage des résultats pour chaque outcome
}
rm(cbcl_vec, outcome, explicative, formula, model, tbl, table_S6_for_outcome)


table_S6 <- list()

for (i in 1:52) {
  tbl <- tbl_merge(
    list(
      tbls_by_outcome_multi$ch_cbclintscore_y2[[i]], # Premier tbl_regression de la première liste de "tbls_by_outcome_multi"
      prep_table_S6$ch_cbclintscore_y2[[i]], # Premier tbl_regression de la première liste de "table_S6"
      tbls_by_outcome_multi$ch_cbclextscore_y2[[i]], # Premier tbl_regression de la deuxième liste de "tbls_by_outcome_multi"
      prep_table_S6$ch_cbclextscore_y2[[i]] # Premier tbl_regression de la deuxième liste de "table_S6"
    ),
    tab_spanner = c("**Internalizing CBCL sub-score at 2 years, adjusted for HOME variable (main analysis)**", 
                    "**Internalizing CBCL sub-score at 2 years, not adjusted for HOME variable**", 
                    "**Externalizing CBCL sub-score at 2 years, adjusted for HOME variable (main analysis)**", 
                    "**Externalizing CBCL sub-score at 2 years, not adjusted for HOME variable**")
  )
  table_S6[[i]] <- tbl   # Ajout de la table fusionnée à la liste
}
table_S6 <- tbl_stack(table_S6)

rm(covariates_sensi_home, i, tbl, prep_table_S6, merge_tbls_function_rich, merge_tbls_function_sha)

## Table S7: Sensitivity analysis – effect of the psy variable on WPSSI Y3 -----
## Sensitivity analysis – Effects of the age of assessment covariate on the WPSSI outcomes assessed at 3 years.

iq_vec_sensi <- c("ch_WPPSI_verbal_comprehension_Y3", "ch_WPPSI_visuospatiale_Y3", "ch_WPPSI_work_memory_Y3", "ch_WPPSI_total_Y3")

# Initialisation de la liste pour stocker les résultats
prep_table_S7 <- list()

for (outcome in iq_vec_sensi) {
  # Initialisation de la sous-liste pour stocker les résultats pour chaque outcome
  table_S7_for_outcome <- list()
  
  for (explicative in explanatory) {
    # Construction de la formule
    formula <- as.formula(paste(outcome, "~", explicative, "+", paste(covariates, collapse = "+")))
    
    # Création du modèle de régression linéaire
    model <- lm(formula, data = bdd_final_imp_1)
    
    tbl <-                                                                      
      tbl_regression(
        model, 
        include = explicative,
        estimate_fun = scales::label_number(accuracy = .01, decimal.mark = "."),
        pvalue_fun = custom_pvalue_fun,
        exponentiate = FALSE) %>%
      bold_p() %>%
      bold_labels() %>%
      add_n()
    
    table_S7_for_outcome[[explicative]] <- tbl        # Stockage du tbl_regression dans la liste
  }
  
  prep_table_S7[[outcome]] <- table_S7_for_outcome    # Stockage des résultats pour chaque outcome
}
rm(iq_vec_sensi, outcome, explicative, formula, model, tbl, table_S7_for_outcome)


table_S7 <- list()

for (i in 1:52) {
  tbl <- tbl_merge(
    list(
      tbls_by_outcome_multi$ch_WPPSI_verbal_comprehension_cor_Y3[[i]], 
      prep_table_S7$ch_WPPSI_verbal_comprehension_Y3[[i]], 
      
      tbls_by_outcome_multi$ch_WPPSI_visuospatiale_cor_Y3[[i]], 
      prep_table_S7$ch_WPPSI_visuospatiale_Y3[[i]], 
      
      tbls_by_outcome_multi$ch_WPPSI_work_memory_cor_Y3[[i]], 
      prep_table_S7$ch_WPPSI_work_memory_Y3[[i]], 
      
      tbls_by_outcome_multi$ch_WPPSI_total_cor_Y3[[i]], 
      prep_table_S7$ch_WPPSI_total_Y3[[i]] 
    ),
    tab_spanner = c("**Verbal comprehension WPPSI sub-score at 3 years, adjusted for neuropsychologist (main analysis)**", 
                    "**Verbal comprehension WPPSI sub-score at 3 years, not adjusted for neuropsychologist**", 
                    
                    "**Visuospatial WPPSI sub-score at 3 years, adjusted for neuropsychologist (main analysis)**", 
                    "**Visuospatial WPPSI sub-score at 3 years, not adjusted for neuropsychologist**", 
                    
                    "**Work memory WPPSI sub-score at 3 years, adjusted for neuropsychologist (main analysis)**", 
                    "**Work memory WPPSI sub-score at 3 years, not adjusted for neuropsychologist**", 
                    
                    "**Total WPPSI score at 3 years, adjusted for neuropsychologist (main analysis)**", 
                    "**Total WPPSI score at 3 years, not adjusted for neuropsychologist**")
  )
  table_S7[[i]] <- tbl   
}
table_S7 <- tbl_stack(table_S7)

rm(tbl, prep_table_S7)


## Table S8: Sensitivity analysis – effect of the predictors of the exposure -----
## Sensitivity analysis – Effects of the age of assessment covariate on the WPSSI outcomes assessed at 3 years.
sensi_explanatory <- explanatory %>% str_replace_all("_std", "")

sensi_covariates <- c("po_w_kg_3cat", 
                      "po_he_3cat", 
                      "mo_dipl_2cat", 
                      "mo_age", 
                      "mo_bmi_bepr_3cat",
                      "ch_sex",
                      "mo_par_2cat",
                      "ch_bf_duration_till48w_4cat",
                      "po_gd",
                      #"po_delmod",
                      "ch_food_intro_Y1_3cat",
                      #"mo_pets",
                      #"ch_antibio_Y1_2cat",
                      "home_total_y3",
                      "mo_hadtotscore_grt3_imp",
                      "mo_tob_gr_anyt_yn_n2",
                      "ch_tabacco_passive_up_to_Y1",
                      "ch_care_main_12m_opt2_2c")

sensi_covariates_CBCL <- c("ch_age_CBCL_Y2", sensi_covariates)     
sensi_ovariates_SRS_BRIEF <- c("ch_age_SRS_BRIEFP_Y3", sensi_covariates)               
sensi_covariates_IQ <- sensi_covariates

sensi_covariates_map <- list(
  CBCL = sensi_covariates_CBCL,
  SRS_BRIEF = sensi_ovariates_SRS_BRIEF,
  IQ = sensi_covariates_IQ)

### standardized on technical factors & not adjusted for the exposure’s predictors ----
prep_table_S8_a <- vector("list", length(outcomes))
names(prep_table_S8_a) <- outcomes

for (outcome in outcomes) {     
  if (outcome %in% c("ch_cbclintscore_y2", "ch_cbclextscore_y2")) {    # Sélection des covariables appropriées
    covars <- sensi_covariates_map$CBCL
  } else if (outcome %in% c("ch_SRStotal_y3", "ch_briefpinhibit_y3", "ch_briefpshift_y3", "ch_briefpemocontrol_y3", "ch_briefpworkmemo_y3", "ch_briefpplan_y3")) {
    covars <- sensi_covariates_map$SRS_BRIEF
  } else if (outcome %in% c("ch_WPPSI_verbal_comprehension_cor_Y3", "ch_WPPSI_visuospatiale_cor_Y3", "ch_WPPSI_work_memory_cor_Y3", "ch_WPPSI_total_cor_Y3")) {
    covars <- sensi_covariates_map$IQ
  }
  
  tbls_for_outcome_multi_sensi_a <- vector("list", length(explanatory))
  names(tbls_for_outcome_multi_sensi_a) <- explanatory
  
  for (exposure in explanatory) {                                               # running linear regression
    terms <- c(exposure, covars)
    formula <- reformulate(terms, response = outcome)
    model <- lm(formula, data = bdd_final_imp_1)
    
    tbl <-                                                                      
      tbl_regression(
        model, 
        include = exposure,
        estimate_fun = scales::label_number(accuracy = .01, decimal.mark = "."),
        pvalue_fun = custom_pvalue_fun,
        exponentiate = FALSE) %>%
      bold_p() %>%
      bold_labels() %>%
      add_global_p(include = exposure, singular.ok = TRUE, keep = TRUE)
    
    tbls_for_outcome_multi_sensi_a[[exposure]] <- tbl
  }
  prep_table_S8_a[[outcome]] <- tbls_for_outcome_multi_sensi_a
}
rm(tbl)

### not standardized on technical factors & not adjusted for the exposure’s predictors ----
prep_table_S8_b <- vector("list", length(outcomes))
names(prep_table_S8_b) <- outcomes

for (outcome in outcomes) {     
  if (outcome %in% c("ch_cbclintscore_y2", "ch_cbclextscore_y2")) {    # Sélection des covariables appropriées
    covars <- sensi_covariates_map$CBCL
  } else if (outcome %in% c("ch_SRStotal_y3", "ch_briefpinhibit_y3", "ch_briefpshift_y3", "ch_briefpemocontrol_y3", "ch_briefpworkmemo_y3", "ch_briefpplan_y3")) {
    covars <- sensi_covariates_map$SRS_BRIEF
  } else if (outcome %in% c("ch_WPPSI_verbal_comprehension_cor_Y3", "ch_WPPSI_visuospatiale_cor_Y3", "ch_WPPSI_work_memory_cor_Y3", "ch_WPPSI_total_cor_Y3")) {
    covars <- sensi_covariates_map$IQ
  }
  
  tbls_for_outcome_multi_sensi_b <- vector("list", length(sensi_explanatory))
  names(tbls_for_outcome_multi_sensi_b) <- sensi_explanatory
  
  for (exposure in sensi_explanatory) {                                            
    terms <- c(exposure, covars)
    formula <- reformulate(terms, response = outcome)
    model <- lm(formula, data = bdd_final_imp_1)
    
    tbl <-                                                                      
      tbl_regression(
        model, 
        include = exposure,
        estimate_fun = scales::label_number(accuracy = .01, decimal.mark = "."),
        pvalue_fun = custom_pvalue_fun,
        exponentiate = FALSE) %>%
      bold_p() %>%
      bold_labels() %>%
      add_global_p(include = exposure, singular.ok = TRUE, keep = TRUE)
    
    tbls_for_outcome_multi_sensi_b[[exposure]] <- tbl
  }
  prep_table_S8_b[[outcome]] <- tbls_for_outcome_multi_sensi_b
}
rm(tbl)

### not standardized on technical factors & adjusted for the exposure’s predictors ----
prep_table_S8_c <- vector("list", length(outcomes))
names(prep_table_S8_c) <- outcomes

for (outcome in outcomes) {     
  if (outcome %in% c("ch_cbclintscore_y2", "ch_cbclextscore_y2")) {    # Sélection des covariables appropriées
    covars <- covariates_map$CBCL
  } else if (outcome %in% c("ch_SRStotal_y3", "ch_briefpinhibit_y3", "ch_briefpshift_y3", "ch_briefpemocontrol_y3", "ch_briefpworkmemo_y3", "ch_briefpplan_y3")) {
    covars <- covariates_map$SRS_BRIEF
  } else if (outcome %in% c("ch_WPPSI_verbal_comprehension_cor_Y3", "ch_WPPSI_visuospatiale_cor_Y3", "ch_WPPSI_work_memory_cor_Y3", "ch_WPPSI_total_cor_Y3")) {
    covars <- covariates_map$IQ
  }
  
  tbls_for_outcome_multi_sensi_c <- vector("list", length(sensi_explanatory))
  names(tbls_for_outcome_multi_sensi_c) <- sensi_explanatory
  
  for (exposure in sensi_explanatory) {                                            
    terms <- c(exposure, covars)
    formula <- reformulate(terms, response = outcome)
    model <- lm(formula, data = bdd_final_imp_1)
    
    tbl <-                                                                      
      tbl_regression(
        model, 
        include = exposure,
        estimate_fun = scales::label_number(accuracy = .01, decimal.mark = "."),
        pvalue_fun = custom_pvalue_fun,
        exponentiate = FALSE) %>%
      bold_p() %>%
      bold_labels() %>%
      add_global_p(include = exposure, singular.ok = TRUE, keep = TRUE)
    
    tbls_for_outcome_multi_sensi_c[[exposure]] <- tbl
  }
  prep_table_S8_c[[outcome]] <- tbls_for_outcome_multi_sensi_c
}
rm(tbl)

### Assemblage avec l'analyse principale
table_S8 <- list()

for (i in 1:52) {
  tbl <- tbl_merge(
    list(
      tbls_by_outcome_multi$ch_cbclintscore_y2[[i]], 
      prep_table_S8_a$ch_cbclintscore_y2[[i]], 
      prep_table_S8_c$ch_cbclintscore_y2[[i]], 
      prep_table_S8_b$ch_cbclintscore_y2[[i]], 
      
      tbls_by_outcome_multi$ch_cbclextscore_y2[[i]], 
      prep_table_S8_a$ch_cbclextscore_y2[[i]], 
      prep_table_S8_c$ch_cbclextscore_y2[[i]], 
      prep_table_S8_b$ch_cbclextscore_y2[[i]], 
      
      tbls_by_outcome_multi$ch_SRStotal_y3[[i]], 
      prep_table_S8_a$ch_SRStotal_y3[[i]],
      prep_table_S8_c$ch_SRStotal_y3[[i]],
      prep_table_S8_b$ch_SRStotal_y3[[i]],
      
      tbls_by_outcome_multi$ch_briefpinhibit_y3[[i]], 
      prep_table_S8_a$ch_briefpinhibit_y3[[i]], 
      prep_table_S8_c$ch_briefpinhibit_y3[[i]], 
      prep_table_S8_b$ch_briefpinhibit_y3[[i]], 
      
      tbls_by_outcome_multi$ch_briefpshift_y3[[i]], 
      prep_table_S8_a$ch_briefpshift_y3[[i]], 
      prep_table_S8_c$ch_briefpshift_y3[[i]], 
      prep_table_S8_b$ch_briefpshift_y3[[i]], 
      
      tbls_by_outcome_multi$ch_briefpemocontrol_y3[[i]], 
      prep_table_S8_a$ch_briefpemocontrol_y3[[i]], 
      prep_table_S8_c$ch_briefpemocontrol_y3[[i]], 
      prep_table_S8_b$ch_briefpemocontrol_y3[[i]], 
      
      tbls_by_outcome_multi$ch_briefpworkmemo_y3[[i]], 
      prep_table_S8_a$ch_briefpworkmemo_y3[[i]], 
      prep_table_S8_c$ch_briefpworkmemo_y3[[i]], 
      prep_table_S8_b$ch_briefpworkmemo_y3[[i]], 
      
      tbls_by_outcome_multi$ch_briefpplan_y3[[i]], 
      prep_table_S8_a$ch_briefpplan_y3[[i]], 
      prep_table_S8_c$ch_briefpplan_y3[[i]], 
      prep_table_S8_b$ch_briefpplan_y3[[i]], 
      
      tbls_by_outcome_multi$ch_WPPSI_verbal_comprehension_cor_Y3[[i]], 
      prep_table_S8_a$ch_WPPSI_verbal_comprehension_cor_Y3[[i]], 
      prep_table_S8_c$ch_WPPSI_verbal_comprehension_cor_Y3[[i]], 
      prep_table_S8_b$ch_WPPSI_verbal_comprehension_cor_Y3[[i]], 
      
      tbls_by_outcome_multi$ch_WPPSI_visuospatiale_cor_Y3[[i]], 
      prep_table_S8_a$ch_WPPSI_visuospatiale_cor_Y3[[i]], 
      prep_table_S8_c$ch_WPPSI_visuospatiale_cor_Y3[[i]], 
      prep_table_S8_b$ch_WPPSI_visuospatiale_cor_Y3[[i]], 
      
      tbls_by_outcome_multi$ch_WPPSI_work_memory_cor_Y3[[i]], 
      prep_table_S8_a$ch_WPPSI_work_memory_cor_Y3[[i]], 
      prep_table_S8_c$ch_WPPSI_work_memory_cor_Y3[[i]], 
      prep_table_S8_b$ch_WPPSI_work_memory_cor_Y3[[i]], 
      
      tbls_by_outcome_multi$ch_WPPSI_total_cor_Y3[[i]], 
      prep_table_S8_a$ch_WPPSI_total_cor_Y3[[i]],
      prep_table_S8_c$ch_WPPSI_total_cor_Y3[[i]],
      prep_table_S8_b$ch_WPPSI_total_cor_Y3[[i]]
    ),
    tab_spanner = c("**Internalizing CBCL sub-score at 2 years, standardized on technical factors & adjusted on exposure’s predictors (main analysis)**", 
                    "**Internalizing CBCL sub-score at 2 years, standardized on technical factors & not adjusted on exposure’s predictors**", 
                    "**Internalizing CBCL sub-score at 2 years, not standardized on technical factors & adjusted on exposure’s predictors**", 
                    "**Internalizing CBCL sub-score at 2 years, not standardized on technical factors & not adjusted on exposure’s predictors**", 
                    
                    "**Externalizing CBCL sub-score at 2 years, standardized on technical factors & adjusted on exposure’s predictors (main analysis)**", 
                    "**Externalizing CBCL sub-score at 2 years, standardized on technical factors & not adjusted on exposure’s predictors**", 
                    "**Externalizing CBCL sub-score at 2 years, not standardized on technical factors & adjusted on exposure’s predictors**", 
                    "**Externalizing CBCL sub-score at 2 years, not standardized on technical factors & not adjusted on exposure’s predictors**", 
                    
                    "**Total SRS score at 3 years, standardized on technical factors & adjusted on exposure’s predictors (main analysis)**", 
                    "**Total SRS score at 3 years, standardized on technical factors & not adjusted on exposure’s predictors**", 
                    "**Total SRS score at 3 years, not standardized on technical factors & adjusted on exposure’s predictors**", 
                    "**Total SRS score at 3 years, not standardized on technical factors & not adjusted on exposure’s predictors**", 
                    
                    "**Inhibition BRIEF-P sub-score at 3 years, standardized on technical factors & adjusted on exposure’s predictors (main analysis)**", 
                    "**Inhibition BRIEF-P sub-score at 3 years, standardized on technical factors & not adjusted on exposure’s predictors**", 
                    "**Inhibition BRIEF-P sub-score at 3 years, not standardized on technical factors & adjusted on exposure’s predictors**", 
                    "**Inhibition BRIEF-P sub-score at 3 years, not standardized on technical factors & not adjusted on exposure’s predictors**", 
                    
                    "**Shift BRIEF-P sub-score at 3 years, standardized on technical factors & adjusted on exposure’s predictors (main analysis)**", 
                    "**Shift BRIEF-P sub-score at 3 years, standardized on technical factors & not adjusted on exposure’s predictors**", 
                    "**Shift BRIEF-P sub-score at 3 years, not standardized on technical factors & adjusted on exposure’s predictors**", 
                    "**Shift BRIEF-P sub-score at 3 years, not standardized on technical factors & not adjusted on exposure’s predictors**", 
                    
                    "**Emotional control BRIEF-P sub-score at 3 years, standardized on technical factors & adjusted on exposure’s predictors (main analysis)**", 
                    "**Emotional control BRIEF-P sub-score at 3 years, standardized on technical factors & not adjusted on exposure’s predictors**", 
                    "**Emotional control BRIEF-P sub-score at 3 years, not standardized on technical factors & adjusted on exposure’s predictors**", 
                    "**Emotional control BRIEF-P sub-score at 3 years, not standardized on technical factors & not adjusted on exposure’s predictors**", 
                    
                    "**Work memory BRIEF-P sub-score at 3 years, standardized on technical factors & adjusted on exposure’s predictors (main analysis)**", 
                    "**Work memory BRIEF-P sub-score at 3 years, standardized on technical factors & not adjusted on exposure’s predictors**", 
                    "**Work memory BRIEF-P sub-score at 3 years, not standardized on technical factors & adjusted on exposure’s predictors**", 
                    "**Work memory BRIEF-P sub-score at 3 years, not standardized on technical factors & not adjusted on exposure’s predictors**", 
                    
                    "**Plan and organization BRIEF-P sub-score at 3 years, standardized on technical factors & adjusted on exposure’s predictors (main analysis)**", 
                    "**Plan and organization BRIEF-P sub-score at 3 years, standardized on technical factors & not adjusted on exposure’s predictors**", 
                    "**Plan and organization BRIEF-P sub-score at 3 years, not standardized on technical factors & adjusted on exposure’s predictors**", 
                    "**Plan and organization BRIEF-P sub-score at 3 years, not standardized on technical factors & not adjusted on exposure’s predictors**", 
                    
                    "**Verbal comprehension WPPSI sub-score at 3 years, standardized on technical factors & adjusted on exposure’s predictors (main analysis)**", 
                    "**Verbal comprehension WPPSI sub-score at 3 years, standardized on technical factors & not adjusted on exposure’s predictors**", 
                    "**Verbal comprehension WPPSI sub-score at 3 years, not standardized on technical factors & adjusted on exposure’s predictors**", 
                    "**Verbal comprehension WPPSI sub-score at 3 years, not standardized on technical factors & not adjusted on exposure’s predictors**", 
                    
                    "**Visuospatial WPPSI sub-score at 3 years, standardized on technical factors & adjusted on exposure’s predictors (main analysis)**", 
                    "**Visuospatial WPPSI sub-score at 3 years, standardized on technical factors & not adjusted on exposure’s predictors**", 
                    "**Visuospatial WPPSI sub-score at 3 years, not standardized on technical factors & adjusted on exposure’s predictors**", 
                    "**Visuospatial WPPSI sub-score at 3 years, not standardized on technical factors & not adjusted on exposure’s predictors**", 
                    
                    "**Work memory WPPSI sub-score at 3 years, standardized on technical factors & adjusted on exposure’s predictors (main analysis)**", 
                    "**Work memory WPPSI sub-score at 3 years, standardized on technical factors & not adjusted on exposure’s predictors**", 
                    "**Work memory WPPSI sub-score at 3 years, not standardized on technical factors & adjusted on exposure’s predictors**", 
                    "**Work memory WPPSI sub-score at 3 years, not standardized on technical factors & not adjusted on exposure’s predictors**", 
                    
                    "**Total WPPSI score at 3 years, standardized on technical factors & adjusted on exposure’s predictors (main analysis)**", 
                    "**Total WPPSI score at 3 years, standardized on technical factors & not adjusted on exposure’s predictors**", 
                    "**Total WPPSI score at 3 years, not standardized on technical factors & adjusted on exposure’s predictors**", 
                    "**Total WPPSI score at 3 years, not standardized on technical factors & not adjusted on exposure’s predictors**")
  )
  table_S8[[i]] <- tbl  
}
table_S8 <- tbl_stack(table_S8)

rm(
  prep_table_S8_a, prep_table_S8_b, prep_table_S8_c,
  terms, formula, model,
  tbl, 
  tbls_for_outcome_multi_sensi_a, tbls_for_outcome_multi_sensi_b, tbls_for_outcome_multi_sensi_c,
  exposure, exposure_name, i, j, outcome, outcome_name, 
  sensi_covariates_CBCL, sensi_covariates_IQ, sensi_ovariates_SRS_BRIEF, sensi_covariates_map, 
  sensi_explanatory)


## Table S9: Sensitivity analysis - non monotonic relation ? ----
## Sensitivity analysis – Non monotonic relation ? heterogeneity and trend tests

# Visualisation des données genres
bdd %>%
  select(contains("ch_feces_rel_g")) %>%
  select(!contains("imp_log")) %>%
  na.omit() %>%
  select_if(~ sum(. != 0, na.rm = TRUE) / length(.) >= 0.66) %>%  # selection des genres ayant une detection >= 66% 
  mutate_all(~ ifelse(.>0, "Yes", "No")) %>%
  tbl_summary(
    type = list(everything() ~ "categorical"))

# Création des vecteurs de variables
genera_to_select <- bdd %>%
  select(contains("ch_feces_rel_g")) %>%
  select(!contains("imp_log")) %>%
  na.omit() %>%
  select_if(~ sum(. != 0, na.rm = TRUE) / length(.) >= 0.66) %>%  # selection des genres ayant une detection >= 66% 
  colnames() %>%
  str_replace_all("_Y1", "_imp_log_std_Y1")

var_to_test <- bdd_final_imp_1 %>%
  select("ch_feces_SpecRich_5000_ASV_std_Y1_10", 
         "ch_feces_Shannon_5000_ASV_std_Y1", 
         "ch_feces_rel_p1_std_Y1_10",
         "ch_feces_rel_p2_std_Y1_10", 
         "ch_feces_rel_p3_std_Y1_10", 
         "ch_feces_rel_p4_std_Y1_10", 
         all_of(genera_to_select)) %>%
  colnames()
var_to_test_ter <- paste(var_to_test, "_ter", sep = "")
var_to_test_ter_num <- paste(var_to_test, "_ter_num", sep = "")
rm(genera_to_select)

# Nettoyage des données 
bdd_hetero <- bdd_final_imp_1 %>%
  select(
    ident, 
    all_of(var_to_test), 
    all_of(outcomes), 
    all_of(covariates)) 

bdd_hetero <- bdd_hetero %>%                                                    # Création des variables tertiles 
  mutate(across(all_of(var_to_test), 
                ~cut(., 
                     breaks = quantile(., probs = 0:3/3, na.rm = TRUE), 
                     labels = c("1st tertile", "2nd tertile", "3rd tertile"), 
                     include.lowest = TRUE),
                .names = "{.col}_ter"))

# bdd_hetero <- bdd_hetero %>%                                                    # Création des variables muettes tertile 2 ou 0 (pour le test hétérogénéité)
#   mutate(across(all_of(var_to_test_ter),
#                 ~ifelse(. == "2nd tertile", 1, 0),
#                 .names = "{.col}_2"))
# 
# bdd_hetero <- bdd_hetero %>%                                                    # Création des variables muettes tertile 3 ou 0 (pour le test hétérogénéité)
#   mutate(across(all_of(var_to_test_ter),
#                 ~ifelse(. == "3rd tertile", 1, 0),
#                 .names = "{.col}_3"))

replace_with_median <- function(data, var, tertile_var) {                       # création des variables numériques à 3 valeurs : les médianes spécifiques des tertiles (pour trend test)
  new_var_name <- paste0(rlang::as_name(ensym(var)), "_ter_num")  
  data %>%
    group_by({{tertile_var}}) %>%
    mutate(!!new_var_name := median({{var}}, na.rm = TRUE)) %>%
    ungroup()
}


bdd_hetero <- bdd_hetero %>%
  replace_with_median(ch_feces_SpecRich_5000_ASV_std_Y1_10, ch_feces_SpecRich_5000_ASV_std_Y1_10_ter) %>%
  replace_with_median(ch_feces_Shannon_5000_ASV_std_Y1, ch_feces_Shannon_5000_ASV_std_Y1_ter) %>%
  replace_with_median(ch_feces_rel_p1_std_Y1_10, ch_feces_rel_p1_std_Y1_10_ter) %>%
  replace_with_median(ch_feces_rel_p2_std_Y1_10, ch_feces_rel_p2_std_Y1_10_ter) %>%    
  replace_with_median(ch_feces_rel_p3_std_Y1_10, ch_feces_rel_p3_std_Y1_10_ter) %>%
  replace_with_median(ch_feces_rel_p4_std_Y1_10, ch_feces_rel_p4_std_Y1_10_ter) %>%
  replace_with_median(ch_feces_rel_g1_imp_log_std_Y1, ch_feces_rel_g1_imp_log_std_Y1_ter) %>%
  replace_with_median(ch_feces_rel_g2_imp_log_std_Y1, ch_feces_rel_g2_imp_log_std_Y1_ter) %>%   
  replace_with_median(ch_feces_rel_g3_imp_log_std_Y1, ch_feces_rel_g3_imp_log_std_Y1_ter) %>%
  replace_with_median(ch_feces_rel_g4_imp_log_std_Y1, ch_feces_rel_g4_imp_log_std_Y1_ter) %>%
  replace_with_median(ch_feces_rel_g5_imp_log_std_Y1, ch_feces_rel_g5_imp_log_std_Y1_ter) %>%
  replace_with_median(ch_feces_rel_g6_imp_log_std_Y1, ch_feces_rel_g6_imp_log_std_Y1_ter) %>%
  replace_with_median(ch_feces_rel_g7_imp_log_std_Y1, ch_feces_rel_g7_imp_log_std_Y1_ter) %>%
  replace_with_median(ch_feces_rel_g8_imp_log_std_Y1, ch_feces_rel_g8_imp_log_std_Y1_ter) %>%
  replace_with_median(ch_feces_rel_g9_imp_log_std_Y1, ch_feces_rel_g9_imp_log_std_Y1_ter) %>%
  replace_with_median(ch_feces_rel_g10_imp_log_std_Y1, ch_feces_rel_g10_imp_log_std_Y1_ter) %>%
  replace_with_median(ch_feces_rel_g13_imp_log_std_Y1, ch_feces_rel_g13_imp_log_std_Y1_ter) %>%
  replace_with_median(ch_feces_rel_g14_imp_log_std_Y1, ch_feces_rel_g14_imp_log_std_Y1_ter) %>%
  replace_with_median(ch_feces_rel_g15_imp_log_std_Y1, ch_feces_rel_g15_imp_log_std_Y1_ter) %>%
  replace_with_median(ch_feces_rel_g19_imp_log_std_Y1, ch_feces_rel_g19_imp_log_std_Y1_ter) %>%    
  replace_with_median(ch_feces_rel_g24_imp_log_std_Y1, ch_feces_rel_g24_imp_log_std_Y1_ter) %>%
  replace_with_median(ch_feces_rel_g25_imp_log_std_Y1, ch_feces_rel_g25_imp_log_std_Y1_ter) %>%
  replace_with_median(ch_feces_rel_g27_imp_log_std_Y1, ch_feces_rel_g27_imp_log_std_Y1_ter) %>%
  replace_with_median(ch_feces_rel_g28_imp_log_std_Y1, ch_feces_rel_g28_imp_log_std_Y1_ter) %>%
  replace_with_median(ch_feces_rel_g29_imp_log_std_Y1, ch_feces_rel_g29_imp_log_std_Y1_ter) %>%
  replace_with_median(ch_feces_rel_g35_imp_log_std_Y1, ch_feces_rel_g35_imp_log_std_Y1_ter) %>%
  replace_with_median(ch_feces_rel_g41_imp_log_std_Y1, ch_feces_rel_g41_imp_log_std_Y1_ter)
rm(replace_with_median)

### Test d'hétérogénéité ----
# ici on met les variables prédicteurs en tertiles avec ajustement sur les covariables
# on fait des régressions linéaires et on regarde les p globales du test de Fisher 
# le test de Fisher compare les modèles avec et sans la variable d'expo en tertile
# la variable d'expo en tertile apporte quelque chose au modèle si p<0.05
# pour le faire sans la fonction add_global_p(test = "F), utiliser le code : anova(test, test_2, test = "F") /!\ attention les n doivent être identiques
fct_table_S9_a <- function(outcome, explicative, data) {
  formule <- as.formula(paste(outcome, "~", explicative,
                              "+ po_w_kg_3cat + 
                              po_he_3cat + 
                              mo_dipl_2cat + 
                              mo_age + 
                              mo_bmi_bepr_3cat + 
                              ch_sex +
                              mo_par_2cat + 
                              ch_bf_duration_till48w_4cat + 
                              po_gd + po_delmod + 
                              ch_food_intro_Y1_3cat + 
                              mo_pets + 
                              ch_antibio_Y1_2cat + 
                              home_total_y3 + 
                              mo_hadtotscore_grt3_imp + 
                              mo_tob_gr_anyt_yn_n2 + 
                              ch_tabacco_passive_up_to_Y1 + 
                              ch_care_main_12m_opt2_2c"))
  modèle <- lm(formule, data = data)
  
  tbl_regression(
    modèle,
    include = !covariates,
    estimate_fun = scales::label_number(accuracy = .01, decimal.mark = "."),
    pvalue_fun = custom_pvalue_fun,
    exponentiate = FALSE) %>%
    bold_p() %>%
    bold_labels() %>%
    add_global_p(test = "F")
}

prep_table_S9_a <- map(outcomes, ~map(var_to_test_ter, fct_table_S9_a, outcome = .x, data = bdd_hetero))
prep_table_S9_a <- set_names(prep_table_S9_a, outcomes)

table_S9_a <- tbl_stack(
  tbls = lapply(1:27, function(j) { 
    tbl_merge(
      tbls = lapply(prep_table_S9_a, function(i) i[[j]]),
      tab_spanner = spanner_names)}))

rm(prep_table_S9_a, fct_table_S9_a)


### Test de tendance ----
# ici on met les variables prédicteurs en variable continues à 3 valeurs correspodant aux médianes spécifiques des tertiles
# on fait des régressions linéaires et on regarde les p de la variable microbiote
# si la p du test d'hétérogénéité est significative, mais pas celle du test de tendance --> sugère une relation non monotome
fct_table_S9_b <- function(outcome, explicative, data) {
  formule <- as.formula(paste(outcome, "~", explicative, 
                              "+ po_w_kg_3cat + 
                              po_he_3cat + 
                              mo_dipl_2cat + 
                              mo_age + 
                              mo_bmi_bepr_3cat + 
                              ch_sex +
                              mo_par_2cat + 
                              ch_bf_duration_till48w_4cat + 
                              po_gd + po_delmod + 
                              ch_food_intro_Y1_3cat + 
                              mo_pets + 
                              ch_antibio_Y1_2cat + 
                              home_total_y3 + 
                              mo_hadtotscore_grt3_imp + 
                              mo_tob_gr_anyt_yn_n2 + 
                              ch_tabacco_passive_up_to_Y1 + 
                              ch_care_main_12m_opt2_2c"))
  modèle <- lm(formule, data = data)
  
  tbl_regression(
    modèle,
    include = !covariates,
    estimate_fun = scales::label_number(accuracy = .01, decimal.mark = "."),
    pvalue_fun = custom_pvalue_fun,
    exponentiate = FALSE) %>%
    bold_p() %>%
    bold_labels() 
}

prep_table_S9_b <- map(outcomes, ~map(var_to_test_ter_num, fct_table_S9_b, outcome = .x, data = bdd_hetero))
prep_table_S9_b <- set_names(prep_table_S9_b, outcomes)

table_S9_b <- tbl_stack(
  tbls = lapply(1:27, function(j) { 
    tbl_merge(
      tbls = lapply(prep_table_S9_b, function(i) i[[j]]),
      tab_spanner = spanner_names)}))

rm(prep_table_S9_b, var_to_test, var_to_test_ter, var_to_test_ter_num, fct_table_S9_b)

table_S9 <- tbl_stack(tbls = list(table_S9_a, table_S9_b))

## Ajout des colonnes Cohen's f pour les tableaux 2, 4 et S4 ----
explanatory_names <- c("Specific richness", "Shannon diversity", 
                       "Firmicutes", "Actinobacteria", "Bacteroidetes", "Proteobacteria", 
                       genera_var_labels)

results <- data.frame()

for (outcome in outcomes) {
  
  if (outcome %in% c("ch_cbclintscore_y2", "ch_cbclextscore_y2")) {
    covariate_specific <- covariates_CBCL
  } else if (outcome %in% c("ch_SRStotal_y3", "ch_briefpinhibit_y3", "ch_briefpshift_y3", "ch_briefpemocontrol_y3", "ch_briefpworkmemo_y3", "ch_briefpplan_y3")) {
    covariate_specific <- covariates_SRS_BRIEF
  } else {
    covariate_specific <- covariates_IQ
  }
  
  for (i in seq_along(explanatory)) {
    
    explanatory_var <- explanatory[i]
    explanatory_name <- explanatory_names[i]
    
    formula_complet <- as.formula(paste(outcome, "~", explanatory_var, "+", paste(covariate_specific, collapse = " + ")))
    formula_reduit <- as.formula(paste(outcome, "~", paste(covariate_specific, collapse = " + ")))
    
    model_complet <- lm(formula_complet, data = bdd_final_imp_1)
    model_reduit <- lm(formula_reduit, data = bdd_final_imp_1)
    
    model_summary <- tidy(model_complet, conf.int = TRUE)
    
    model_summary_exp <- model_summary %>%
      filter(term == explanatory_var) %>%
      select(term, estimate, conf.low, conf.high, p.value) %>%
      mutate(outcome = outcome)
    
    R2_complet <- summary(model_complet)$r.squared
    R2_reduit <- summary(model_reduit)$r.squared
    
    Cohens_f2 <-  (R2_complet - R2_reduit) / (1 - R2_complet)
    
    model_summary_exp <- model_summary_exp %>%
      mutate(Cohens_f2, term_2 = explanatory_name)
    
    results <- rbind(results, model_summary_exp)
  }
}
rm(outcome, covariate_specific, explanatory_var, explanatory_name, model_complet, model_reduit, model_summary, model_summary_exp, i)

results <- results %>%
  mutate(
    Beta = format(round(estimate, 2), digits = 2),
    conf.low = format(round(conf.low, 2), digits = 2),
    conf.high = format(round(conf.high, 2), digits = 2),
    "95% CI" = paste(conf.low, conf.high, sep = ","), 
    p.value = case_when(p.value < 0.001 ~ format(round(p.value, 4), nsmall = 4),
                        p.value < 0.01 ~ format(round(p.value, 3), nsmall = 3),
                        p.value > 0.01 ~ format(round(p.value, 2), nsmall = 2)), 
    Cohens_f2 = case_when(Cohens_f2 <0 ~ "<0",
                          Cohens_f2> 0 & Cohens_f2 < 0.001 ~ "<0.001",
                          Cohens_f2> 0 & Cohens_f2 < 0.01 ~ format(round(Cohens_f2, 3), nsmall = 3),
                          Cohens_f2> 0 & Cohens_f2 > 0.01 ~ format(round(Cohens_f2, 2), nsmall = 2), 
                          .default = as.character(Cohens_f2))) %>%
  select(outcome, term, term_2, Beta, `95% CI`, `p.value`, Cohens_f2)

table_2_bis <- results %>%
  filter(term_2 %in% c("Specific richness", "Shannon diversity")) %>%
  rename(Explanatory = term_2) %>%
  select(-term)

table_4_bis <- results %>%
  filter(term_2 %in% c("Firmicutes", "Actinobacteria", "Bacteroidetes", "Proteobacteria")) %>%
  rename(Explanatory = term_2) %>%
  select(-term)

table_S4_bis <- results %>%
  filter(!term_2 %in% c("Specific richness", "Shannon diversity", 
                        "Firmicutes", "Actinobacteria", 
                        "Bacteroidetes", "Proteobacteria")) %>%
  select(-term) %>%
  pivot_wider(names_from = outcome, 
              values_from = c("Beta", "95% CI", "p.value", "Cohens_f2")) %>%
  mutate(
    term_2 = fct_recode(term_2, 
                        "Clostridium IV" = "Clostridium_IV",
                        "Clostridium sensu stricto" = "Clostridium_sensu_stricto",
                        "Clostridium XlVa" = "Clostridium_XlVa",
                        "Clostridium XVIII" = "Clostridium_XVIII",
                        "Erysipelotrichaceae incertae sedis" = "Erysipelotrichaceae_incertae_sedis",
                        "Escherichia and Shigella" = "Escherichia_Shigella",
                        "Lachnospiracea incertae sedis" = "Lachnospiracea_incertae_sedis",
                        "Ruminococcus 2" = "Ruminococcus2",
                        "Saccharibacteria genera incertae sedis" = "Saccharibacteria_genera_incertae_sedis"))%>%
  select("Gut microbiota parameters" = term_2, 
         contains("ch_cbclintscore_y2"), 
         contains("ch_cbclextscore_y2"),
         contains("ch_SRStotal_y3"), 
         contains("ch_briefpinhibit_y3"),
         contains("ch_briefpshift_y3"), 
         contains("ch_briefpemocontrol_y3"), 
         contains("ch_briefpworkmemo_y3"), 
         contains("ch_briefpplan_y3"), 
         contains("ch_WPPSI_verbal_comprehension_cor_Y3"), 
         contains("ch_WPPSI_visuospatiale_cor_Y3"), 
         contains("ch_WPPSI_work_memory_cor_Y3"), 
         contains("ch_WPPSI_total_cor_Y3"))

write.xlsx(table_2_bis, file = "4_output/table_2_bis.xlsx")
write.xlsx(table_4_bis, file = "4_output/table_4_bis.xlsx")
write.xlsx(table_S4_bis, file = "4_output/table_S4_bis.xlsx")


table_2_bis %>% filter(p.value<0.05 & Cohens_f2>0.02)
table_4_bis %>% filter(p.value<0.05 & Cohens_f2>0.02)
table_S4_bis %>% filter(p.value<0.05 & Cohens_f2>0.02)
results %>% filter(p.value<0.05 & Cohens_f2>0.019) %>% View()
results %>% filter(p.value<0.05 & Cohens_f2>0.12) %>% View()




# Investifation effet confinement covid ----
bdd <- bdd %>%
  mutate(
    confinement_SRS_BRIEF_P = 
      ifelse((cy3hat1_date_creation >= as.Date("2020-03-17") & cy3hat1_date_creation <= as.Date("2020-07-11")) |  # dates du 1er confinement + jusqu'à 2 mois après
               (cy3hat1_date_creation >= as.Date("2020-10-30") & cy3hat1_date_creation <= as.Date("2021-02-15"))|  # dates du 2eme confinement + jusqu'à 2 mois après 
               (cy3hat1_date_creation >= as.Date("2021-04-03") & cy3hat1_date_creation <= as.Date("2021-07-03")), # dates du 3eme confinement + jusqu'à 2 mois après 
             "Yes", "No"), 
    confinement_CBCL = 
      ifelse((cy2ebc1_date_creation >= as.Date("2020-03-17") & cy2ebc1_date_creation <= as.Date("2020-07-11")) |  # dates du 1er confinement + jusqu'à 2 mois après
               (cy2ebc1_date_creation >= as.Date("2020-10-30") & cy2ebc1_date_creation <= as.Date("2021-02-15"))|  # dates du 2eme confinement + jusqu'à 2 mois après 
               (cy3hat1_date_creation >= as.Date("2021-04-03") & cy3hat1_date_creation <= as.Date("2021-07-03")), # dates du 3eme confinement + jusqu'à 2 mois après 
             "Yes", "No"))

bdd %>% select(confinement_CBCL, cy2ebc1_date_creation) %>% arrange(cy2ebc1_date_creation) %>% View()
bdd %>% select(confinement_SRS_BRIEF_P, cy3hat1_date_creation) %>% arrange(cy3hat1_date_creation)%>% View()

bdd %>% 
  filter(!is.na(ch_feces_rel_ASV1035_Y1)) %>%
  select(confinement_CBCL, confinement_SRS_BRIEF_P) %>% 
  tbl_summary(
    type = list(confinement_CBCL ~ "categorical", 
                confinement_SRS_BRIEF_P ~ "categorical"))

bdd_filtered <- bdd %>% filter(!is.na(ch_feces_rel_ASV1035_Y1)) 

tbl_merge(
  tbls = list(
    lm(ch_SRStotal_y3 ~ confinement_SRS_BRIEF_P, data = bdd_filtered) %>% tbl_regression(type = confinement_SRS_BRIEF_P ~ "categorical"),
    lm(ch_briefpinhibit_y3 ~ confinement_SRS_BRIEF_P, data = bdd_filtered) %>% tbl_regression(type = confinement_SRS_BRIEF_P ~ "categorical"), 
    lm(ch_briefpshift_y3 ~ confinement_SRS_BRIEF_P, data = bdd_filtered) %>% tbl_regression(type = confinement_SRS_BRIEF_P ~ "categorical"), 
    lm(ch_briefpemocontrol_y3 ~ confinement_SRS_BRIEF_P, data = bdd_filtered) %>% tbl_regression(type = confinement_SRS_BRIEF_P ~ "categorical"), 
    lm(ch_briefpworkmemo_y3 ~ confinement_SRS_BRIEF_P, data = bdd_filtered) %>% tbl_regression(type = confinement_SRS_BRIEF_P ~ "categorical"), 
    lm(ch_briefpplan_y3 ~ confinement_SRS_BRIEF_P, data = bdd_filtered) %>% tbl_regression(type = confinement_SRS_BRIEF_P ~ "categorical")), 
  tab_spanner = c("ch_SRStotal_y3",
                  "ch_briefpinhibit_y3",
                  "ch_briefpshift_y3",
                  "ch_briefpemocontrol_y3",
                  "ch_briefpworkmemo_y3",
                  "ch_briefpplan_y3"))


summary(test)
bdd %>% 
  filter(!is.na(ch_feces_rel_ASV1035_Y1)) %>%
  select(confinement_CBCL, confinement_SRS_BRIEF_P) %>% 
  tbl_regression(
    lm(ch_SRStotal_y3 ~ confinement_SRS_BRIEF_P, data = bdd), 
    type = list(confinement_CBCL ~ "categorical", 
                confinement_SRS_BRIEF_P ~ "categorical"))





# Additional figures ----
## Figure S1: DAG ----
## Direct acyclic graph of the relation between one year child gut microbiota and neurodevelopment. 


## Figure S2: correlations neurodevelopmental ----
## xx correlations between neurodevelopmental parameters (n between X and X).
figure_S2 <- bdd %>% 
  filter(!is.na(ch_feces_rel_p1_Y1)) %>%
  select(all_of(outcomes)) 
colnames(figure_S2) <- spanner_names
names(figure_S2) <- gsub(" sub-score at 2 years", " Y2", names(figure_S2))
names(figure_S2) <- gsub(" sub-score at 3 years", " Y3", names(figure_S2))
names(figure_S2) <- gsub(" score at 2 years", " Y2", names(figure_S2))
names(figure_S2) <- gsub(" score at 3 years", " Y3", names(figure_S2))
names(figure_S2) <- gsub("SRS", "SRS-II", names(figure_S2))
names(figure_S2) <- gsub("WPPSI", "WPPSI-IV", names(figure_S2))

# permet de savoir les pairwise.complete.obs 
apply(combn(names(figure_S2), 2), 2, function(x) sum(!is.na(figure_S2[[x[1]]]) & !is.na(figure_S2[[x[2]]])))


figure_S2 <- cor(figure_S2, 
                    use = "pairwise.complete.obs", 
                    method = "pearson")

plot.new()
tiff(filename = "4_output/figures/Fig.S2 heatmap_cor_neuro.tiff", units = "mm", width = 250, height = 250, res = 300)
corrplot(figure_S2, 
         method = 'color', 
         type = "lower", 
         tl.col = 'black', 
         tl.srt = 45, 
         addCoef.col = "black",
         number.cex = 1,
         number.digits = 1,
         tl.cex = 1,
         col = rev(COL2(diverging = "RdYlBu")))
dev.off()



## Figure SX: correlations gut microbiota parameters ----
## xx correlations between gut microbiota parameters (n between X and X). 
figure_SX <- bdd %>% 
  select(all_of(genera_var_names)) 
colnames(figure_S2) <- genera_var_labels
figure_S2 <- figure_S2 %>%
  rename(
    "Escherichia and Shigella" = Escherichia_Shigella,
    "Clostridium XlVa" = Clostridium_XlVa, 
    "Lachnospiracea incertae sedis" = Lachnospiracea_incertae_sedis, 
    "Clostridium XVIII" = Clostridium_XVIII, 
    "Clostridium sensu stricto" = Clostridium_sensu_stricto, 
    "Ruminococcus 2" = Ruminococcus2, 
    "Clostridium IV" = Clostridium_IV, 
    "Erysipelotrichaceae incertae sedis" = Erysipelotrichaceae_incertae_sedis, 
    "Saccharibacteria genera incertae sedis" = Saccharibacteria_genera_incertae_sedis)

figure_SX <- cor(figure_SX, 
                 use = "pairwise.complete.obs", 
                 method = "pearson")

plot.new()
tiff(filename = "4_output/Fig.SX heatmap_cor_genera.tiff", units = "mm", width = 250, height = 250, res = 300)
corrplot(figure_SX, 
         method = 'color', 
         type = "lower", 
         tl.col = 'black', 
         tl.srt = 45, 
         # addCoef.col = "black",
         # number.cex = 0.5,
         # number.digits = 1,
         tl.cex = 0.5,
         col = rev(COL2(diverging = "RdYlBu")))
dev.off()



## Figure SX: correlations gut microbiota and covariates ----
## xx correlations between gut microbiota parameters and covariates (n between X and X).
figure_S2 <- bdd_final_imp_1 %>%
  select(all_of(explanatory),
         po_w_kg, 
         po_he, 
         mo_age, 
         mo_par,
         mo_bmi_bepr, 
         po_gd, 
         home_total_y3, 
         mo_hadtotscore_grt3_imp) %>%
  select_if(is.numeric)

# permet de savoir les pairwise.complete.obs 
colSums(!is.na(figure_S2))

figure_S2 <- round(cor(figure_S2,
                       use = "pairwise.complete.obs",
                       method = "pearson"), 1)

figure_S2 <- figure_S2 %>%
  as.data.frame() %>%
  select(all_of(explanatory)) %>%
  t() %>%
  as.data.frame() %>%
  select(!all_of(explanatory)) %>%
  as.matrix()

colnames(figure_S2) <- colnames(figure_S2) %>%
  str_replace_all(
    c("po_w_kg" = "Birth weight (kg)",
      "po_he" = "Birth length (cm)",
      "mo_age" = "Mat. age before preg.",
      "mo_par" = "Mat. parity",
      "mo_bmi_bepr" = "Mat. BMI before preg.",
      "po_gd" = "Gestational age (weeks)",
      "home_total_y3" = "HOME score at 3 years old",
      "mo_hadtotscore_grt3_imp" = "Mat. anxiety and dep, 3rd trim. of preg."))


rownames(figure_S2) <-rownames(figure_S2) %>%
  str_replace_all(
    c("ch_feces_SpecRich_5000_ASV_Y1_10" = "Specific richness",
      "ch_feces_Shannon_5000_ASV_Y1" = "Shannon diversity",
      "ch_feces_rel_p1_Y1_10" = "Firmicutes",
      "ch_feces_rel_p2_Y1_10" = "Actinobacteria",
      "ch_feces_rel_p3_Y1_10" = "Bacteroidetes",
      "ch_feces_rel_p4_Y1_10" = "Proteobacteria",
      "Escherichia_Shigella" = "Escherichia and Shigella", 
      "Clostridium_XlVa" = "Clostridium XlVa",
      "Lachnospiracea_incertae_sedis" = "Lachnospiracea incertae sedis",
      "Clostridium_XVIII" = "Clostridium XVIII",
      "Clostridium_sensu_stricto" = "Clostridium sensu stricto",
      "Ruminococcus2" = "Ruminococcus 2", 
      "Clostridium_IV" = "Clostridium IV", 
      "Erysipelotrichaceae_incertae_sedis" = "Erysipelotrichaceae incertae sedis", 
      "Saccharibacteria_genera_incertae_sedis" = "Saccharibacteria genera incertae sedis"))


tiff(filename = "4_output/Fig.S2 heatmap_cor_neuro_covar.tiff", units = "mm", width = 250, height = 600, res = 300)
corrplot(figure_S2,
         method = 'color',
         tl.col = 'black',
         tl.srt = 55,
         addCoef.col = "black",
         number.cex = 1,
         tl.cex = 1,
         number.digits = 1,
         col = rev(COL2(diverging = "RdYlBu")))
dev.off()


rm(covariates_CBCL, covariates_IQ, covariates_SRS_BRIEF, covariates_map)

## Figure SX: Forestplot alpha div ----
figure_1 <- table_multi %>% 
  filter(`Gut microbiota parameters` %in% c("Shannon diversity",
                                            "Specific richness")) %>%
  mutate(Beta = as.numeric(Beta), 
         Outcome = 
           fct_relevel(Outcome,
                       "Total WPPSI score at 3 years", "Work memory WPPSI sub-score at 3 years", 
                       "Visuospatial WPPSI sub-score at 3 years","Verbal comprehension WPPSI sub-score at 3 years", 
                       "Plan and organization BRIEF-P sub-score at 3 years", "Work memory BRIEF-P sub-score at 3 years",
                       "Emotional control BRIEF-P sub-score at 3 years","Shift BRIEF-P sub-score at 3 years",
                       "Inhibition BRIEF-P sub-score at 3 years", "Total SRS score at 3 years",
                       "Externalizing CBCL sub-score at 2 years", "Internalizing CBCL sub-score at 2 years"), 
         `Gut microbiota parameters` = 
           fct_relevel(`Gut microbiota parameters`, 
                       "Specific richness", "Shannon diversity")) %>%
  ggplot(aes(x = Outcome, 
             y = Beta, 
             min = lower_CI, 
             ymax = upper_CI, 
             color = p_value_shape)) +
  geom_hline(yintercept = 0, linetype="dashed") +
  geom_pointrange(
    position = position_dodge(width = 0.5), 
    size = 0.4) +
  labs(x = "Neurodevelopement", y = "", color = "p-value") +
  theme_lucid() +
  coord_flip()  +
  facet_wrap(~`Gut microbiota parameters`, scales = "free_x", ncol = 2) +
  theme(
    legend.position = "right",
    legend.box = "vertical", 
    legend.justification = "right") +
  scale_color_manual(values = c("p-value<0.05" = "red", "p-value≥0.05" = "black"))

figure_1
ggsave("4_output/fig.1 forest_plot_alphadiv.tiff", 
       figure_1, 
       device = "tiff",
       units = "cm",
       dpi = 300,
       height = 15, 
       width = 25)

## Figure SX: Graphique modèle dose réponse ----

# Vizualisation of the significant results ----
table_multi %>% 
  filter(`p-value` <0.05) %>% 
  filter(`Gut microbiota parameters` != "Proteobacteria") %>% 
  select(-sens_beta, -p_value_shape, -`q-value`, -q_value_shape) %>% 
  select(Phyla_corres, Class_corres, Order_corres, Family_corres, `Gut microbiota parameters`, everything()) %>% 
  arrange(Class_corres, Order_corres, Family_corres) %>% 
  View()

table_multi %>% 
  filter(`p-value` <0.05) %>% 
  filter(`Gut microbiota parameters` %in% c("Firmicutes", "Actinobacteria", "Bacteroidetes", "Proteobacteria")) %>% 
  select(`Gut microbiota parameters`, Outcome, Beta, `95% CI`, lower_CI, upper_CI, `p-value`) %>% 
  View()


table_multi %>% 
  filter(`p-value` <0.05) %>% 
  filter(improved_neuro == "Improved neurodevelopmental outcome") %>%
  filter(!`Gut microbiota parameters` %in% c("Firmicutes", "Actinobacteria", "Bacteroidetes", "Proteobacteria")) %>% 
  select(-sens_beta, -p_value_shape, -`q-value`, -q_value_shape, -Outcome, -`95% CI`, -lower_CI, -upper_CI) %>% 
  select(Phyla_corres, Class_corres, Order_corres, Family_corres, `Gut microbiota parameters`, Exposure, everything()) %>% 
  arrange(`p-value`) %>% 
  View()

table_multi %>% 
    filter(`p-value` <0.05) %>% 
  filter(improved_neuro == "Altered neurodevelopmental outcome") %>%
  filter(!`Gut microbiota parameters` %in% c("Firmicutes", "Actinobacteria", "Bacteroidetes", "Proteobacteria")) %>% 
  select(-sens_beta, -p_value_shape, -`q-value`, -q_value_shape, -Outcome, -`95% CI`, -lower_CI, -upper_CI) %>% 
  select(Phyla_corres, Class_corres, Order_corres, Family_corres, `Gut microbiota parameters`, Exposure, everything()) %>% 
  arrange(`p-value`) %>% 
  View()


# Exporter les résultats ----
save.image("4_output/results_alpha_taxa_2024_04_28.RData")

# Réimporter les résultats ----
load("4_output/results_alpha_taxa_2024_04_28.RData")
